@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "description": "domready held back.",
   "scripts": {
-    "test": "jest",
+    "test": "jest --silent",
     "build": "gulp build",
     "dev": "gulp dev"
   },
@@ -16,48 +16,55 @@
     "setupFiles": [
       "../../../jest/environment.js"
     ],
-    "unmockedModulePathPatterns": [
-      "node_modules/react",
-      "node_modules/underscore",
-      "libs/react-patches",
-      "jsdom"
-    ]
+    "setupTestFrameworkScriptFile": "../../../node_modules/jest-enzyme/lib/index.js"
   },
   "devDependencies": {
-    "babel-core": "^6.7.7",
-    "babel-jest": "^12.0.1",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-transform-object-rest-spread": "^6.6.5",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "css-loader": "^0.23.1",
-    "eslint": "^3.12.2",
-    "eslint-config-airbnb": "^13.0.0",
+    "babel-core": "^6.21.0",
+    "babel-helper-builder-react-jsx": "^6.22.0",
+    "babel-helpers": "^6.22.0",
+    "babel-jest": "^18.0.0",
+    "babel-loader": "^6.2.10",
+    "babel-plugin-transform-object-rest-spread": "^6.20.2",
+    "babel-plugin-transform-react-display-name": "^6.22.0",
+    "babel-plugin-transform-react-jsx": "^6.22.0",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.16.0",
+    "babel-register": "^6.22.0",
+    "css-loader": "^0.26.1",
+    "enzyme": "^2.8.1",
+    "eslint": "^3.13.0",
+    "eslint-config-airbnb": "^14.0.0",
+    "eslint-config-airbnb-base": "^11.0.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.8.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
+    "eslint-plugin-react": "^6.9.0",
+    "espree": "^3.3.2",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
     "gulp": "^3.9.1",
     "gulp-util": "^3.0.7",
-    "jest-cli": "^12.0.1",
+    "jest": "^19.0.2",
+    "jest-enzyme": "^2.1.2",
     "jsdom": "^8.4.0",
     "json-loader": "^0.5.4",
     "node-sass": "^3.8.0",
-    "sass-loader": "^4.0.0",
-    "string-replace-loader": "^1.0.0",
+    "react-addons-test-utils": "^15.5.1",
+    "react-test-renderer": "^15.5.4",
+    "sass-loader": "^4.1.1",
+    "string-replace-loader": "^1.0.5",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.0"
+    "webpack": "^1.14.0"
   },
   "dependencies": {
     "auth0-lock": "^10.4.0",
-    "babel-polyfill": "^6.7.4",
+    "babel-polyfill": "^6.20.0",
     "brace": "^0.3.0",
     "chart.js": "^2.5.0",
     "ckeditor": "file:node_shims/ckeditor",
     "color": "^0.10.1",
     "cookie-monster": "0.0.7",
+    "create-react-class": "^15.5.2",
     "d3": "^3.4.6",
     "dagre-d3": "git+https://github.com/ENCODE-DCC/dagre-d3.git#v1.2.0",
     "dalliance": "git://github.com/ENCODE-DCC/dalliance.git#commonjs-require-13.x",
@@ -68,12 +75,15 @@
     "jsonschema": "^1.1.0",
     "marked": "^0.3.3",
     "moment": "^2.17.1",
+    "prop-types": "^15.5.6",
     "query-string": "^4.1.0",
-    "react": "^0.12.2",
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3",
     "scriptjs": "^2.5.7",
-    "source-map-support": "^0.2.7",
+    "shortid": "^2.2.6",
+    "source-map-support": "^0.4.11",
     "subprocess-middleware": "^0.1.0",
     "underscore": "^1.8.3",
-    "whatwg-fetch": "git+https://github.com/lrowe/fetch.git#bf5d58b738fb2ed6d60791b944d36075fee8604a"
+    "whatwg-fetch": "git+https://github.com/ENCODE-DCC/fetch.git"
   }
 }
