@@ -1,12 +1,14 @@
 const React = require('react');
+import PropTypes from 'prop-types';
+import createReactClass from 'create-react-class';
 const form = require('../form');
 const globals = require('../globals');
 const item = require('../item');
 
-const FallbackBlockView = React.createClass({
+const FallbackBlockView = createReactClass({
     propTypes: {
-        blocktype: React.PropTypes.object,
-        value: React.PropTypes.any,
+        blocktype: PropTypes.object,
+        value: PropTypes.any,
     },
 
     render() {
@@ -20,11 +22,11 @@ const FallbackBlockView = React.createClass({
     },
 });
 
-const JSONInput = React.createClass({
+const JSONInput = createReactClass({
 
     propTypes: {
-        value: React.PropTypes.any,
-        onChange: React.PropTypes.func,
+        value: PropTypes.any,
+        onChange: PropTypes.func,
     },
 
     getInitialState() {
@@ -60,11 +62,11 @@ const fallbackSchema = {
     formInput: <JSONInput />,
 };
 
-module.exports.FallbackBlockEdit = React.createClass({
+module.exports.FallbackBlockEdit = createReactClass({
     propTypes: {
-        schema: React.PropTypes.object,
-        value: React.PropTypes.any,
-        onChange: React.PropTypes.func,
+        schema: PropTypes.object,
+        value: PropTypes.any,
+        onChange: PropTypes.func,
     },
 
     update(name, value) {
