@@ -1,5 +1,7 @@
 'use strict';
 var React = require('react');
+import PropTypes from 'prop-types';
+import createReactClass from 'create-react-class';
 var fetched = require('../fetched');
 var globals = require('../globals');
 var item = require('./item');
@@ -11,7 +13,7 @@ var ObjectPicker = require('../inputs').ObjectPicker;
 var RichTextBlockView = richtext.RichTextBlockView;
 
 
-var TeaserCore = React.createClass({
+var TeaserCore = createReactClass({
     renderImage: function() {
         var context = this.props.value.image;
         if (typeof context === 'object') {
@@ -41,7 +43,7 @@ var TeaserCore = React.createClass({
 });
 
 
-var TeaserBlockView = React.createClass({
+var TeaserBlockView = createReactClass({
 
     getDefaultProps: function() {
         return {value: {
@@ -73,10 +75,10 @@ var TeaserBlockView = React.createClass({
 });
 
 
-var RichEditor = React.createClass({
+var RichEditor = createReactClass({
 
     childContextTypes: {
-        editable: React.PropTypes.bool
+        editable: PropTypes.bool
     },
 
     getChildContext: function() {
