@@ -1,5 +1,7 @@
 'use strict';
 var React = require('react');
+import PropTypes from 'prop-types';
+import createReactClass from 'create-react-class';
 var _ = require('underscore');
 var url = require('url');
 var panel = require('../libs/bootstrap/panel');
@@ -36,10 +38,10 @@ const EXCERPT_LENGTH = 80; // Maximum number of characters in an excerpt
 //     }
 // ]
 
-var DocumentsPanel = module.exports.DocumentsPanel = React.createClass({
+var DocumentsPanel = module.exports.DocumentsPanel = createReactClass({
     propTypes: {
-        documentSpecs: React.PropTypes.array.isRequired, // List of document arrays and their titles
-        title: React.PropTypes.string // Title of the document panel
+        documentSpecs: PropTypes.array.isRequired, // List of document arrays and their titles
+        title: PropTypes.string // Title of the document panel
     },
 
     render: function() {
@@ -82,9 +84,9 @@ var DocumentsPanel = module.exports.DocumentsPanel = React.createClass({
 });
 
 
-var DocumentsSubpanels = module.exports.DocumentsSubpanels = React.createClass({
+var DocumentsSubpanels = module.exports.DocumentsSubpanels = createReactClass({
     propTypes: {
-        documentSpec: React.PropTypes.object.isRequired // List of document arrays and their titles
+        documentSpec: PropTypes.object.isRequired // List of document arrays and their titles
     },
 
     render: function() {
@@ -106,7 +108,7 @@ var DocumentsSubpanels = module.exports.DocumentsSubpanels = React.createClass({
 
 // Display a single document within a <DocumentPanel>. This routine requires that you register display components for each
 // of the five major parts of a single document panel. See globals.js for a guide to the parts.
-var Document = module.exports.Document = React.createClass({
+var Document = module.exports.Document = createReactClass({
     getInitialState: function() {
         return {panelOpen: false};
     },
@@ -149,9 +151,9 @@ var Document = module.exports.Document = React.createClass({
 
 
 // Document header component -- default
-var DocumentHeader = module.exports.DocumentHeader = React.createClass({
+var DocumentHeader = module.exports.DocumentHeader = createReactClass({
     propTypes: {
-        doc: React.PropTypes.object.isRequired // Document object to render
+        doc: PropTypes.object.isRequired // Document object to render
     },
 
     render: function() {
@@ -166,9 +168,9 @@ var DocumentHeader = module.exports.DocumentHeader = React.createClass({
 });
 
 // Document caption component -- default
-var DocumentCaption = module.exports.DocumentCaption = React.createClass({
+var DocumentCaption = module.exports.DocumentCaption = createReactClass({
     propTypes: {
-        doc: React.PropTypes.object.isRequired // Document object to render
+        doc: PropTypes.object.isRequired // Document object to render
     },
 
     render: function() {
@@ -192,9 +194,9 @@ var DocumentCaption = module.exports.DocumentCaption = React.createClass({
 });
 
 // Document preview component -- default
-var DocumentPreview = module.exports.DocumentPreview = React.createClass({
+var DocumentPreview = module.exports.DocumentPreview = createReactClass({
     propTypes: {
-        doc: React.PropTypes.object.isRequired // Document object to render
+        doc: PropTypes.object.isRequired // Document object to render
     },
 
     render: function() {
@@ -208,11 +210,11 @@ var DocumentPreview = module.exports.DocumentPreview = React.createClass({
 
 
 // Document file component -- default
-var DocumentFile = module.exports.DocumentFile = React.createClass({
+var DocumentFile = module.exports.DocumentFile = createReactClass({
     propTypes: {
-        doc: React.PropTypes.object.isRequired, // Document object to render
-        detailOpen: React.PropTypes.bool, // True if detail panel is visible
-        detailSwitch: React.PropTypes.func // Parent component function to call when detail switch clicked
+        doc: PropTypes.object.isRequired, // Document object to render
+        detailOpen: PropTypes.bool, // True if detail panel is visible
+        detailSwitch: PropTypes.func // Parent component function to call when detail switch clicked
     },
 
     render: function() {
@@ -249,11 +251,11 @@ var DocumentFile = module.exports.DocumentFile = React.createClass({
 
 
 // Document detail component -- default
-var DocumentDetail = module.exports.DocumentDetail = React.createClass({
+var DocumentDetail = module.exports.DocumentDetail = createReactClass({
     propTypes: {
-        doc: React.PropTypes.object.isRequired, // Document object to render
-        detailOpen: React.PropTypes.bool, // True if detail panel is visible
-        key: React.PropTypes.string // Unique key for identification
+        doc: PropTypes.object.isRequired, // Document object to render
+        detailOpen: PropTypes.bool, // True if detail panel is visible
+        key: PropTypes.string // Unique key for identification
     },
 
     render: function() {
@@ -263,7 +265,7 @@ var DocumentDetail = module.exports.DocumentDetail = React.createClass({
 
         return (
             <div className={keyClass}>
-                <dl className='key-value-doc' id={'panel-' + this.props.id} aria-labeledby={'tab-' + this.props.id} role="tabpanel">
+                <dl className='key-value-doc' id={'panel-' + this.props.id} aria-labelledby={'tab-' + this.props.id} role="tabpanel">
                     {excerpt ?
                         <div data-test="caption">
                             <dt>Description</dt>
@@ -314,9 +316,9 @@ globals.document_views.file.register(DocumentFile, 'Document');
 globals.document_views.detail.register(DocumentDetail, 'Document');
 
 
-const QCAttachmentCaption = React.createClass({
+const QCAttachmentCaption = createReactClass({
     propTypes: {
-        title: React.PropTypes.string.isRequired, // Title to display for attachment
+        title: PropTypes.string.isRequired, // Title to display for attachment
     },
 
     render: function () {
@@ -334,10 +336,10 @@ const QCAttachmentCaption = React.createClass({
 });
 
 
-const QCAttachmentPreview = React.createClass({
+const QCAttachmentPreview = createReactClass({
     propTypes: {
-        context: React.PropTypes.object.isRequired, // QC metric object that owns the attachment to render
-        attachment: React.PropTypes.object.isRequired, // Attachment to render
+        context: PropTypes.object.isRequired, // QC metric object that owns the attachment to render
+        attachment: PropTypes.object.isRequired, // Attachment to render
     },
 
     render: function () {
@@ -352,12 +354,12 @@ const QCAttachmentPreview = React.createClass({
 });
 
 // Display a panel for attachments that aren't a part of an associated document
-const AttachmentPanel = module.exports.AttachmentPanel = React.createClass({
+const AttachmentPanel = module.exports.AttachmentPanel = createReactClass({
     propTypes: {
-        context: React.PropTypes.object.isRequired, // Object that owns the attachment; needed for attachment path
-        attachment: React.PropTypes.object.isRequired, // Attachment being rendered
-        title: React.PropTypes.string, // Title to display in the caption area
-        modal: React.PropTypes.bool, // `true` if attachments are displayed in a modal
+        context: PropTypes.object.isRequired, // Object that owns the attachment; needed for attachment path
+        attachment: PropTypes.object.isRequired, // Attachment being rendered
+        title: PropTypes.string, // Title to display in the caption area
+        modal: PropTypes.bool, // `true` if attachments are displayed in a modal
     },
 
     render: function () {
