@@ -1,19 +1,22 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import createReactClass from 'create-react-class';
 import _ from 'underscore';
+import url from 'url';
 import { Panel, PanelHeading, PanelBody } from '../libs/bootstrap/panel';
 import { collapseIcon } from '../libs/svg-icons';
+import { auditDecor } from './audit';
 import { SortTable } from './sorttable';
 import globals from './globals';
 import StatusLabel from './statuslabel';
 import { ProjectBadge, Attachment } from './image';
-import { AuditIndicators, AuditDetail, AuditMixin } from './audit';
 import { RelatedItems } from './item';
 import { DbxrefList } from './dbxref';
 import { Breadcrumbs } from './navigation';
 import { treatmentDisplay, singleTreatment } from './objectutils';
 import { BiosampleTable } from './typeutils';
-import { DocumentsPanel, AttachmentPanel } from './doc';
-import { PickerActionsMixin } from './search';
+import { DocumentsPanel } from './doc';
+import { PickerActions } from './search';
 
 
 // Map GM techniques to a presentable string
@@ -23,9 +26,9 @@ const GM_TECHNIQUE_MAP = {
 };
 
 
-const GeneticModificationCharacterizations = React.createClass({
+const GeneticModificationCharacterizations = createReactClass({
     propTypes: {
-        characterizations: React.PropTypes.array, // Genetic modificiation characterizations to display
+        characterizations: PropTypes.array.isRequired, // Genetic modificiation characterizations to display
     },
 
     render: function () {
@@ -62,12 +65,12 @@ function geneticModificationTechniques(techniques) {
 }
 
 
-export const GeneticModification = React.createClass({
+export const GeneticModificationComponent = createReactClass({
     propTypes: {
-        context: React.PropTypes.object, // GM object being displayed
+        context: PropTypes.object.isRequired, // GM object being displayed
+        auditIndicators: PropTypes.func.isRequired, // Audit HOC function to display audit indicators
+        auditDetail: PropTypes.func.isRequired, // Audit HOC function to display audit details
     },
-
-    mixins: [AuditMixin],
 
     render: function () {
         const context = this.props.context;
@@ -133,11 +136,11 @@ export const GeneticModification = React.createClass({
                             <div className="characterization-status-labels">
                                 <StatusLabel title="Status" status={context.status} />
                             </div>
-                            <AuditIndicators audits={context.audit} id="genetic-modification-audit" />
+                            {this.props.auditIndicators(context.audit, 'genetic-modification-audit')}
                         </div>
                     </div>
                 </header>
-                <AuditDetail audits={context.audit} except={context['@id']} id="genetic-modification-audit" />
+                {this.props.auditDetail(context.audit, 'genetic-modification-audit', { except: context['@id'] })}
                 <Panel addClasses="data-display">
                     <PanelBody addClasses="panel-body-with-header">
                         <div className="flexrow">
@@ -296,12 +299,14 @@ export const GeneticModification = React.createClass({
     },
 });
 
+const GeneticModification = auditDecor(GeneticModificationComponent);
+
 globals.content_views.register(GeneticModification, 'GeneticModification');
 
 
-const GMAttachmentCaption = React.createClass({
+const GMAttachmentCaption = createReactClass({
     propTypes: {
-        title: React.PropTypes.string.isRequired, // Title to display for attachment
+        title: PropTypes.string.isRequired, // Title to display for attachment
     },
 
     render: function () {
@@ -319,10 +324,10 @@ const GMAttachmentCaption = React.createClass({
 });
 
 
-const GMAttachmentPreview = React.createClass({
+const GMAttachmentPreview = createReactClass({
     propTypes: {
-        context: React.PropTypes.object.isRequired, // QC metric object that owns the attachment to render
-        attachment: React.PropTypes.object.isRequired, // Attachment to render
+        context: PropTypes.object.isRequired, // QC metric object that owns the attachment to render
+        attachment: PropTypes.object.isRequired, // Attachment to render
     },
 
     render: function () {
@@ -345,9 +350,9 @@ globals.document_views.preview.register(GMAttachmentPreview, 'GeneticModificatio
 
 
 // Display modification technique specific to the CRISPR type.
-const TechniqueCrispr = React.createClass({
+const TechniqueCrispr = createReactClass({
     propTypes: {
-        context: React.PropTypes.object, // CRISPR genetic modificiation technique to display
+        context: PropTypes.object.isRequired, // CRISPR genetic modificiation technique to display
     },
 
     render: function () {
@@ -408,9 +413,9 @@ globals.panel_views.register(TechniqueCrispr, 'Crispr');
 
 
 // Display modification technique specific to the TALE type.
-const TechniqueTale = React.createClass({
+const TechniqueTale = createReactClass({
     propTypes: {
-        context: React.PropTypes.object, // TALE genetic modificiation technique to display
+        context: PropTypes.object.isRequired, // TALE genetic modificiation technique to display
     },
 
     render: function () {
@@ -473,12 +478,68 @@ const TechniqueTale = React.createClass({
 globals.panel_views.register(TechniqueTale, 'Tale');
 
 
-const Listing = React.createClass({
+// Display a panel for attachments that aren't a part of an associated document
+export const AttachmentPanel = createReactClass({
     propTypes: {
-        context: React.PropTypes.object, // Search results object
+        context: PropTypes.object.isRequired, // Object that owns the attachment; needed for attachment path
+        attachment: PropTypes.object.isRequired, // Attachment being rendered
+        title: PropTypes.string, // Title to display in the caption area
     },
 
-    mixins: [PickerActionsMixin, AuditMixin],
+    getDefaultProps: function () {
+        return {
+            title: '',
+        };
+    },
+
+    render: function () {
+        const { context, attachment, title } = this.props;
+
+        // Make the download link
+        let download;
+        let attachmentHref;
+        if (attachment.href && attachment.download) {
+            attachmentHref = url.resolve(context['@id'], attachment.href);
+            download = (
+                <div className="dl-link">
+                    <i className="icon icon-download" />&nbsp;
+                    <a data-bypass="true" href={attachmentHref} download={attachment.download}>
+                        Download
+                    </a>
+                </div>
+            );
+        } else {
+            download = <em>Attachment not available to download</em>;
+        }
+
+        return (
+            <div className="flexcol panel-attachment">
+                <Panel addClasses={globals.itemClass(context, 'view-detail')}>
+                    <figure>
+                        <Attachment context={context} attachment={attachment} className="characterization" />
+                    </figure>
+                    <div className="document-intro document-meta-data">
+                        {title ?
+                            <div data-test="attachments">
+                                <strong>Method: </strong>
+                                {title}
+                            </div>
+                        : null}
+                        {download}
+                    </div>
+                </Panel>
+            </div>
+        );
+    },
+});
+
+
+const ListingComponent = createReactClass({
+    propTypes: {
+        context: PropTypes.object.isRequired, // Search results object
+        auditDetail: PropTypes.func.isRequired, // Audit HOC function to show audit details
+        auditIndicators: PropTypes.func.isRequired, // Audit HOC function to display audit indicators
+    },
 
     render: function () {
         const result = this.props.context;
@@ -499,22 +560,24 @@ const Listing = React.createClass({
         return (
             <li>
                 <div className="clearfix">
-                    {this.renderActions()}
+                    <PickerActions {...this.props} />
                     <div className="pull-right search-meta">
                         <p className="type meta-title">Genetic modifications</p>
                         <p className="type meta-status">{` ${result.status}`}</p>
-                        <AuditIndicators audits={result.audit} id={result['@id']} search />
+                        {this.props.auditIndicators(result.audit, result['@id'], { search: true })}
                     </div>
                     <div className="accession"><a href={result['@id']}>{result.modification_type}</a></div>
                     <div className="data-row">
                         {techniques.length ? <div><strong>Modification techniques: </strong>{techniques.join(', ')}</div> : null}
                     </div>
                 </div>
-                <AuditDetail audits={result.audit} except={result['@id']} id={result['@id']} forcedEditLink />
+                {this.props.auditDetail(result.audit, result['@id'], { except: result['@id'], forcedEditLink: true })}
             </li>
         );
     },
 });
+
+const Listing = auditDecor(ListingComponent);
 
 globals.listing_views.register(Listing, 'GeneticModification');
 
@@ -576,9 +639,9 @@ export const calcGMSummarySentence = _.memoize(rCalcGMSummarySentence, gm => gm.
 // Display a summary of genetic modifications given in the geneticModifications prop. This
 // component assumes the `geneticModifications` array has at least one entry, so make sure of that
 // before calling this component.
-export const GeneticModificationSummary = React.createClass({
+export const GeneticModificationSummary = createReactClass({
     propTypes: {
-        geneticModifications: React.PropTypes.array.isRequired, // Array of genetic modifications
+        geneticModifications: PropTypes.array.isRequired, // Array of genetic modifications
     },
 
     render: function () {
@@ -626,10 +689,10 @@ export const GeneticModificationSummary = React.createClass({
 
 // Display one GM group, which consists of all GMs that share the same type, technique, target, and
 // treatments. A group is an array of GM objects.
-export const GeneticModificationGroup = React.createClass({
+export const GeneticModificationGroup = createReactClass({
     propTypes: {
-        groupSentence: React.PropTypes.string, // GM group detail sentence to display
-        gms: React.PropTypes.array, // GM objects to display within a group
+        groupSentence: PropTypes.string.isRequired, // GM group detail sentence to display
+        gms: PropTypes.array.isRequired, // GM objects to display within a group
     },
 
     getInitialState: function () {
