@@ -1,106 +1,98 @@
-'use strict';
-var React = require('react');
-var _ = require('underscore');
-var moment = require('moment');
-var globals = require('./globals');
-var {FetchedData, FetchedItems, Param} = require('./fetched');
-import NewsPreviews from './news';
-var cloneWithProps = require('react/lib/cloneWithProps');
-var panel = require('../libs/bootstrap/panel');
-var {Panel, PanelBody, PanelHeading} = panel;
+import React from 'react';
+import PropTypes from 'prop-types';
+import _ from 'underscore';
+import moment from 'moment';
+import { FetchedData, FetchedItems, Param } from './fetched';
+import { Panel, PanelBody } from '../libs/bootstrap/panel';
 
 
 const newsUri = '/search/?type=Page&news=true&status=released';
 
 
-// Main page component to render the home page
-var Home = module.exports.Home = React.createClass({
+// Convert the selected organisms and assays into an encoded query.
+function generateQuery(selectedOrganisms, selectedAssayCategory) {
+    // Make the base query.
+    let query = selectedAssayCategory === 'COMPPRED' ? '?type=Annotation&encyclopedia_version=3' : '?type=Experiment&status=released';
 
-    getInitialState: function () { // sets initial state for current and newtabs
-        return {
-            current: "?type=Experiment&status=released", // show all released experiments
-            organisms: [], // create empty array of selected tabs
-            assayCategory: "",
-            socialHeight: 0
+    // Add the selected assay category, if any (doesn't apply to Computational Predictions).
+    if (selectedAssayCategory && selectedAssayCategory !== 'COMPPRED') {
+        query += `&assay_slims=${selectedAssayCategory}`;
+    }
+
+    // Add all the selected organisms, if any
+    if (selectedOrganisms.length) {
+        const organismSpec = selectedAssayCategory === 'COMPPRED' ? 'organism.scientific_name=' : 'replicates.library.biosample.donor.organism.scientific_name=';
+        const queryStrings = {
+            HUMAN: `${organismSpec}Homo+sapiens`, // human
+            MOUSE: `${organismSpec}Mus+musculus`, // mouse
+            WORM: `${organismSpec}Caenorhabditis+elegans`, // worm
+            FLY: `${organismSpec}Drosophila+melanogaster&${organismSpec}Drosophila+pseudoobscura&${organismSpec}Drosophila+simulans&${organismSpec}Drosophila+mojavensis&${organismSpec}Drosophila+ananassae&${organismSpec}Drosophila+virilis&${organismSpec}Drosophila+yakuba`,
         };
-    },
+        const organismQueries = selectedOrganisms.map(organism => queryStrings[organism]);
+        query += `&${organismQueries.join('&')}`;
+    }
 
-    handleAssayCategoryClick: function (assayCategory) {
+    return query;
+}
+
+
+// Main page component to render the home page
+export default class Home extends React.Component {
+    constructor(props) {
+        super(props);
+
+        // Set initial React state.
+        this.state = {
+            current: '?type=Experiment&status=released', // show all released experiments
+            organisms: [], // create empty array of selected tabs
+            assayCategory: '',
+            socialHeight: 0,
+        };
+
+        // Required binding of `this` to component methods or else they can't see `this`.
+        this.handleAssayCategoryClick = this.handleAssayCategoryClick.bind(this);
+        this.handleTabClick = this.handleTabClick.bind(this);
+        this.newsLoaded = this.newsLoaded.bind(this);
+    }
+
+    handleAssayCategoryClick(assayCategory) {
         if (this.state.assayCategory === assayCategory) {
             this.setState({ assayCategory: '' });
         } else {
             this.setState({ assayCategory: assayCategory });
         }
-    },
+    }
 
-    // pass in string with organism query and either adds or removes tab from list of selected tabs
-    handleTabClick: function (selectedTab) {
-
-        var tempArray = _.clone(this.state.organisms); // creates a copy of this.state.newtabs
-
-        if (tempArray.indexOf(selectedTab) == -1) {
+    handleTabClick(selectedTab) {
+        // Create a copy of this.state.newtabs so we can manipulate it in peace.
+        const tempArray = _.clone(this.state.organisms);
+        if (tempArray.indexOf(selectedTab) === -1) {
             // if tab isn't already in array, then add it
             tempArray.push(selectedTab);
         } else {
             // otherwise if it is in array, remove it from array and from link
-            let indexToRemoveArray = tempArray.indexOf(selectedTab);
+            const indexToRemoveArray = tempArray.indexOf(selectedTab);
             tempArray.splice(indexToRemoveArray, 1);
         }
 
-        // update newtabs
+        // Update the list of user-selected organisms.
         this.setState({ organisms: tempArray });
-    },
+    }
 
-    newsLoaded: function () {
-        // Called once the news content gets loaded
-        let newsEl = this.refs.newslisting.getDOMNode();
-        this.setState({ socialHeight: newsEl.clientHeight });
-    },
+    // Called when the news content loads so that we can get its height. That lets us match up the
+    // height of <TwitterWidget>. If we don't have any news items, nodeRef has `undefined` and we
+    // just hard-code the height at 600 so that the Twitter widget has some space.
+    newsLoaded(nodeRef) {
+        this.setState({ socialHeight: nodeRef ? nodeRef.clientHeight : 600 });
+    }
 
-    // Convert the selected organisms and assays into an encoded query.
-    generateQuery: function (selectedOrganisms, selectedAssayCategory) {
-        // Make the base query
-        let query = selectedAssayCategory === 'COMPPRED' ? '?type=Annotation&encyclopedia_version=3' : "?type=Experiment&status=released";
-
-        // Add the selected assay category, if any (doesn't apply to Computational Predictions)
-        if (selectedAssayCategory && selectedAssayCategory !== 'COMPPRED') {
-            query += '&assay_slims=' + selectedAssayCategory;
-        }
-
-        // Add all the selected organisms, if any
-        if (selectedOrganisms.length) {
-            let organismSpec = selectedAssayCategory === 'COMPPRED' ? 'organism.scientific_name=' : 'replicates.library.biosample.donor.organism.scientific_name=';
-            let queryStrings = {
-                'HUMAN': organismSpec + 'Homo+sapiens', // human
-                'MOUSE': organismSpec + 'Mus+musculus', // mouse
-                'WORM': organismSpec + 'Caenorhabditis+elegans', // worm
-                'FLY': organismSpec + 'Drosophila+melanogaster&' + // fly
-                organismSpec + 'Drosophila+pseudoobscura&' +
-                organismSpec + 'Drosophila+simulans&' +
-                organismSpec + 'Drosophila+mojavensis&' +
-                organismSpec + 'Drosophila+ananassae&' +
-                organismSpec + 'Drosophila+virilis&' +
-                organismSpec + 'Drosophila+yakuba'
-            };
-            let organismQueries = selectedOrganisms.map(organism => queryStrings[organism]);
-            query += '&' + organismQueries.join('&');
-        }
-
-        return query;
-    },
-
-    render: function () {
+    render() {
         // Based on the currently selected organisms and assay category, generate a query string
         // for the GET request to retrieve chart data.
-        let currentQuery = this.generateQuery(this.state.organisms, this.state.assayCategory);
+        const currentQuery = generateQuery(this.state.organisms, this.state.assayCategory);
 
         return (
             <div className="whole-page">
-                <header className="row">
-                    <div className="col-sm-12">
-                        <h1 className="page-title"></h1>
-                    </div>
-                </header>
                 <div className="row">
                     <div className="col-xs-12">
                         <Panel>
@@ -118,7 +110,7 @@ var Home = module.exports.Home = React.createClass({
                                     <div className="news-header">
                                         <h2>News <a href="/news/" title="More ENCODE news" className="twitter-ref">More ENCODE news</a></h2>
                                     </div>
-                                    <NewsLoader ref="newslisting" newsLoaded={this.newsLoaded} />
+                                    <NewsLoader newsLoaded={this.newsLoaded} />
                                 </div>
                                 <div className="social-twitter">
                                     <TwitterWidget height={this.state.socialHeight} />
@@ -130,47 +122,48 @@ var Home = module.exports.Home = React.createClass({
             </div>
         );
     }
-
-});
+}
 
 
 // Given retrieved data, draw all home-page charts.
-var ChartGallery = React.createClass({
-    render: function () {
-        return (
-            <PanelBody>
-                <div className="view-all">
-                    <a href={"/matrix/" + this.props.query} className="view-all-button btn btn-info btn-sm" role="button">View Assay Matrix</a>
-                </div>
-                <div className="chart-gallery">
-                    <div className="chart-single">
-                        <HomepageChart {...this.props} />
-                    </div>
-                    <div className="chart-single">
-                        <HomepageChart2 {...this.props} />
-                    </div>
-                    <div className="chart-single">
-                        <HomepageChart3 {...this.props} />
-                    </div>
-                </div>
-            </PanelBody>
-        );
-    }
-});
+const ChartGallery = props => (
+    <PanelBody>
+        <div className="view-all">
+            <a href={`/matrix/${props.query}`} className="view-all-button btn btn-info btn-sm" role="button">View Assay Matrix</a>
+        </div>
+        <div className="chart-gallery">
+            <div className="chart-single">
+                <HomepageChart {...props} />
+            </div>
+            <div className="chart-single">
+                <HomepageChart2 {...props} />
+            </div>
+            <div className="chart-single">
+                <HomepageChart3 {...props} />
+            </div>
+        </div>
+    </PanelBody>
+);
+
+ChartGallery.propTypes = {
+    query: PropTypes.string, // Query string to add to /matrix/ URI
+};
 
 
 // Component to allow clicking boxes on classic image
-var AssayClicking = React.createClass({
-    propTypes: {
-        assayCategory: React.PropTypes.string
-    },
+class AssayClicking extends React.Component {
+    constructor(props) {
+        super(props);
+
+        // Required binding of `this` to component methods or else they can't see `this`.
+        this.sortByAssay = this.sortByAssay.bind(this);
+    }
 
     // Properly adds or removes assay category from link
-    sortByAssay: function (category, e) {
-
-        function handleClick(category, ctx) {
+    sortByAssay(category, e) {
+        function handleClick(cat, ctx) {
             // Call the Home component's function to record the new assay cateogry
-            ctx.props.handleAssayCategoryClick(category); // handles assay category click
+            ctx.props.handleAssayCategoryClick(cat); // handles assay category click
         }
 
         if (e.type === 'touchend') {
@@ -181,18 +174,21 @@ var AssayClicking = React.createClass({
         } else {
             this.assayClickHandled = false;
         }
-    },
+    }
 
     // Renders classic image and svg rectangles
-    render: function () {
-        const assayList = ["3D+chromatin+structure",
-            "DNA+accessibility",
-            "DNA+binding",
-            "DNA+methylation",
-            "COMPPRED",
-            "Transcription",
-            "RNA+binding"];
-        let assayCategory = this.props.assayCategory;
+    render() {
+        const assayList = [
+            '3D+chromatin+structure',
+            'DNA+accessibility',
+            'DNA+binding',
+            'DNA+methylation',
+            'COMPPRED',
+            'Transcription',
+            'RNA+binding',
+        ];
+        const assayCategory = this.props.assayCategory;
+
         return (
             <div ref="graphdisplay">
                 <div className="overall-classic">
@@ -201,16 +197,16 @@ var AssayClicking = React.createClass({
 
                     <div className="site-banner">
                         <div className="site-banner-img">
-                            <img src="static/img/classic-image.jpg" />
+                            <img src="static/img/classic-image.jpg" alt="ENCODE representational diagram with embedded assay selection buttons" />
 
                             <svg id="site-banner-overlay" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 2260 1450" className="classic-svg">
-                                <rect id={assayList[0]} x="101.03" y="645.8" width="257.47" height="230.95" className={"rectangle-box" + (assayCategory == assayList[0] ? " selected" : "")} onClick={this.sortByAssay.bind(null, assayList[0])} onTouchEnd={this.sortByAssay.bind(null, assayList[0])} />
-                                <rect id={assayList[1]} x="386.6" y="645.8" width="276.06" height="230.95" className={"rectangle-box" + (assayCategory == assayList[1] ? " selected" : "")} onClick={this.sortByAssay.bind(null, assayList[1])} onTouchEnd={this.sortByAssay.bind(null, assayList[1])} />
-                                <rect id={assayList[2]} x="688.7" y="645.8" width="237.33" height="230.95" className={"rectangle-box" + (assayCategory == assayList[2] ? " selected" : "")} onClick={this.sortByAssay.bind(null, assayList[2])} onTouchEnd={this.sortByAssay.bind(null, assayList[2])} />
-                                <rect id={assayList[3]} x="950.83" y="645.8" width="294.65" height="230.95" className={"rectangle-box" + (assayCategory == assayList[3] ? " selected" : "")} onClick={this.sortByAssay.bind(null, assayList[3])} onTouchEnd={this.sortByAssay.bind(null, assayList[3])} />
-                                <rect id={assayList[4]} x="1273.07" y="645.8" width="373.37" height="230.95" className={"rectangle-box" + (assayCategory == assayList[4] ? " selected" : "")} onClick={this.sortByAssay.bind(null, assayList[4])} onTouchEnd={this.sortByAssay.bind(null, assayList[4])} />
-                                <rect id={assayList[5]} x="1674.06" y="645.8" width="236.05" height="230.95" className={"rectangle-box" + (assayCategory == assayList[5] ? " selected" : "")} onClick={this.sortByAssay.bind(null, assayList[5])} onTouchEnd={this.sortByAssay.bind(null, assayList[5])} />
-                                <rect id={assayList[6]} x="1937.74" y="645.8" width="227.38" height="230.95" className={"rectangle-box" + (assayCategory == assayList[6] ? " selected" : "")} onClick={this.sortByAssay.bind(null, assayList[6])} onTouchEnd={this.sortByAssay.bind(null, assayList[6])} />
+                                <BannerOverlayButton item={assayList[0]} x="101.03" y="645.8" width="257.47" height="230.95" selected={assayCategory === assayList[0]} clickHandler={this.sortByAssay} />
+                                <BannerOverlayButton item={assayList[1]} x="386.6" y="645.8" width="276.06" height="230.95" selected={assayCategory === assayList[1]} clickHandler={this.sortByAssay} />
+                                <BannerOverlayButton item={assayList[2]} x="688.7" y="645.8" width="237.33" height="230.95" selected={assayCategory === assayList[2]} clickHandler={this.sortByAssay} />
+                                <BannerOverlayButton item={assayList[3]} x="950.83" y="645.8" width="294.65" height="230.95" selected={assayCategory === assayList[3]} clickHandler={this.sortByAssay} />
+                                <BannerOverlayButton item={assayList[4]} x="1273.07" y="645.8" width="373.37" height="230.95" selected={assayCategory === assayList[4]} clickHandler={this.sortByAssay} />
+                                <BannerOverlayButton item={assayList[5]} x="1674.06" y="645.8" width="236.05" height="230.95" selected={assayCategory === assayList[5]} clickHandler={this.sortByAssay} />
+                                <BannerOverlayButton item={assayList[6]} x="1937.74" y="645.8" width="227.38" height="230.95" selected={assayCategory === assayList[6]} clickHandler={this.sortByAssay} />
                             </svg>
                         </div>
 
@@ -227,70 +223,119 @@ var AssayClicking = React.createClass({
             </div>
         );
     }
+}
 
-});
+AssayClicking.propTypes = {
+    assayCategory: PropTypes.string.isRequired, // Test to display in each audit's detail, possibly containing @ids that this component turns into links automatically
+};
+
+
+// Draw an overlay button on the ENCODE banner.
+const BannerOverlayButton = (props) => {
+    const { item, x, y, width, height, selected, clickHandler } = props;
+
+    return (
+        <rect
+            id={item}
+            x={x}
+            y={y}
+            width={width}
+            height={height}
+            className={`rectangle-box${selected ? ' selected' : ''}`}
+            onClick={(e) => { clickHandler(item, e); }}
+        />
+    );
+};
+
+BannerOverlayButton.propTypes = {
+    item: PropTypes.string, // ID of button being clicked
+    x: PropTypes.string, // X coordinate of button
+    y: PropTypes.string, // Y coordinate of button
+    width: PropTypes.string, // Width of button in pixels
+    height: PropTypes.string, // Height of button in pixels
+    selected: PropTypes.bool, // `true` if button is selected
+    clickHandler: PropTypes.func, // Function to call when the button is clicked
+};
+
 
 // Passes in tab to handleTabClick
-var TabClicking = React.createClass({
-    propTypes: {
-        organisms: React.PropTypes.array, // Array of currently selected tabs
-        handleTabClick: React.PropTypes.func
-    },
-
-    render: function () {
-        let organisms = this.props.organisms;
+class TabClicking extends React.Component {
+    render() {
+        const { organisms, handleTabClick } = this.props;
         return (
             <div ref="tabdisplay">
                 <div className="organism-selector">
-                    <a className={"single-tab" + (organisms.indexOf('HUMAN') != -1 ? " selected" : "")} href="#" data-trigger onClick={this.props.handleTabClick.bind(null, 'HUMAN')}>Human</a>
-                    <a className={"single-tab" + (organisms.indexOf('MOUSE') != -1 ? " selected" : "")} href="#" data-trigger onClick={this.props.handleTabClick.bind(null, 'MOUSE')}>Mouse</a>
-                    <a className={"single-tab" + (organisms.indexOf('WORM') != -1 ? " selected" : "")} href="#" data-trigger onClick={this.props.handleTabClick.bind(null, 'WORM')}>Worm</a>
-                    <a className={"single-tab" + (organisms.indexOf('FLY') != -1 ? " selected" : "")} href="#" data-trigger onClick={this.props.handleTabClick.bind(null, 'FLY')}>Fly</a>
+                    <OrganismSelector organism="Human" selected={organisms.indexOf('HUMAN') !== -1} clickHandler={handleTabClick} />
+                    <OrganismSelector organism="Mouse" selected={organisms.indexOf('MOUSE') !== -1} clickHandler={handleTabClick} />
+                    <OrganismSelector organism="Worm" selected={organisms.indexOf('WORM') !== -1} clickHandler={handleTabClick} />
+                    <OrganismSelector organism="Fly" selected={organisms.indexOf('FLY') !== -1} clickHandler={handleTabClick} />
                 </div>
             </div>
         );
     }
+}
 
-});
+TabClicking.propTypes = {
+    organisms: PropTypes.array, // Array of currently selected tabs
+    handleTabClick: PropTypes.func, // Function to call when a tab is clicked
+};
+
+
+const OrganismSelector = (props) => {
+    const { organism, selected, clickHandler } = props;
+
+    return (
+        <button className={`organism-selector__tab${selected ? ' organism-selector--selected' : ''}`} onClick={() => { clickHandler(organism.toUpperCase(organism)); }}>
+            {organism}
+        </button>
+    );
+};
+
+OrganismSelector.propTypes = {
+    organism: PropTypes.string, // Organism this selector represents
+    selected: PropTypes.bool, // `true` if selector is selected
+    clickHandler: PropTypes.func, // Function to call to handle a selector click
+};
 
 
 // Initiates the GET request to search for experiments, and then pass the data to the HomepageChart
 // component to draw the resulting chart.
-var HomepageChartLoader = React.createClass({
-    propTypes: {
-        query: React.PropTypes.string // Current search URI based on selected assayCategory
-    },
+const HomepageChartLoader = (props) => {
+    const { query, organisms, assayCategory } = props;
 
-    render: function () {
-        return (
-            <FetchedData ignoreErrors>
-                <Param name="data" url={'/search/' + this.props.query} />
-                <ChartGallery organisms={this.props.organisms} assayCategory={this.props.assayCategory} query={this.props.query} />
-            </FetchedData>
-        );
-    }
+    return (
+        <FetchedData ignoreErrors>
+            <Param name="data" url={`/search/${query}`} />
+            <ChartGallery organisms={organisms} assayCategory={assayCategory} query={query} />
+        </FetchedData>
+    );
+};
 
-});
+HomepageChartLoader.propTypes = {
+    query: PropTypes.string, // Current search URI based on selected assayCategory
+    organisms: PropTypes.array, // Array of selected organism strings
+    assayCategory: PropTypes.string, // Selected assay category
+};
 
 
 // Draw the total chart count in the middle of the donut.
 function drawDonutCenter(chart) {
-    let canvasId = chart.chart.canvas.id;
-    let width = chart.chart.width;
-    let height = chart.chart.height;
-    let ctx = chart.chart.ctx;
+    const canvasId = chart.chart.canvas.id;
+    const width = chart.chart.width;
+    const height = chart.chart.height;
+    const ctx = chart.chart.ctx;
 
     ctx.fillStyle = '#000000';
     ctx.restore();
-    let fontSize = (height / 114).toFixed(2);
-    ctx.font = fontSize + "em sans-serif";
-    ctx.textBaseline = "middle";
+    const fontSize = (height / 114).toFixed(2);
+    ctx.font = `${fontSize}em sans-serif`;
+    ctx.textBaseline = 'middle';
 
     if (canvasId === 'myChart' || canvasId === 'myChart2') {
-        let data = chart.data.datasets[0].data;
-        let total = data.reduce((prev, curr) => prev + curr);
-        let textX = Math.round((width - ctx.measureText(total).width) / 2);
-        let textY = height / 2;
+        const data = chart.data.datasets[0].data;
+        const total = data.reduce((prev, curr) => prev + curr);
+        const textX = Math.round((width - ctx.measureText(total).width) / 2);
+        const textY = height / 2;
 
         ctx.clearRect(0, 0, width, height);
         ctx.fillText(total, textX, textY);
@@ -302,25 +347,41 @@ function drawDonutCenter(chart) {
 
 
 // Component to display the D3-based chart for Project
-let HomepageChart = React.createClass({
+class HomepageChart extends React.Component {
+    constructor(props) {
+        super(props);
+        this.wrapperHeight = 200;
+        this.createChart = this.createChart.bind(this);
+        this.updateChart = this.updateChart.bind(this);
+    }
 
-    contextTypes: {
-        navigate: React.PropTypes.func,
-        projectColors: React.PropTypes.object // DataColor instance for experiment project
-    },
+    componentDidMount() {
+        // Create the chart, and assign the chart to this.myPieChart when the process finishes.
+        if (document.getElementById('myChart')) {
+            this.createChart(this.facetData);
+        }
+    }
 
-    wrapperHeight: 200,
+    componentDidUpdate() {
+        if (this.myPieChart) {
+            // Existing data updated
+            this.updateChart(this.myPieChart, this.facetData);
+        } else if (this.facetData.length) {
+            // Chart existed but was destroyed for lack of data. Rebuild the chart.
+            this.createChart(this.facetData);
+        }
+    }
 
     // Draw the Project chart, for initial load, or when the previous load had no data for this
     // chart.
-    createChart: function (facetData) {
-        require.ensure(['chart.js'], function (require) {
-            let Chart = require('chart.js');
+    createChart(facetData) {
+        require.ensure(['chart.js'], (require) => {
+            const Chart = require('chart.js');
 
             // for each item, set doc count, add to total doc count, add proper label, and assign color.
-            let colors = this.context.projectColors.colorList(facetData.map(term => term.key), { shade: 10 });
-            let data = [];
-            let labels = [];
+            const colors = this.context.projectColors.colorList(facetData.map(term => term.key), { shade: 10 });
+            const data = [];
+            const labels = [];
 
             // Convert facet data to chart data.
             facetData.forEach((term, i) => {
@@ -331,77 +392,79 @@ let HomepageChart = React.createClass({
             // adding total doc count to middle of donut
             // http://stackoverflow.com/questions/20966817/how-to-add-text-inside-the-doughnut-chart-using-chart-js/24671908
             Chart.pluginService.register({
-                beforeDraw: drawDonutCenter
+                beforeDraw: drawDonutCenter,
             });
 
             // Pass the assay_title counts to the charting library to render it.
-            var canvas = document.getElementById("myChart");
-            var ctx = canvas.getContext("2d");
-            this.myPieChart = new Chart(ctx, {
-                type: 'doughnut',
-                data: {
-                    labels: labels,
-                    datasets: [{
-                        data: data,
-                        backgroundColor: colors
-                    }]
-                },
+            const canvas = document.getElementById('myChart');
+            if (canvas) {
+                const ctx = canvas.getContext('2d');
+                this.myPieChart = new Chart(ctx, {
+                    type: 'doughnut',
+                    data: {
+                        labels: labels,
+                        datasets: [{
+                            data: data,
+                            backgroundColor: colors,
+                        }],
+                    },
 
-                options: {
-                    responsive: true,
-                    maintainAspectRatio: false,
-                    legend: {
-                        display: false // Hide automatically generated legend; we draw it ourselves
-                    },
-                    animation: {
-                        duration: 200
-                    },
-                    legendCallback: function (chart) { // allows for legend clicking
-                        let data = chart.data.datasets[0].data;
-                        let text = [];
-                        text.push('<ul>');
-                        for (let i = 0; i < data.length; i++) {
-                            if (data[i]) {
-                                text.push('<li>');
-                                text.push('<a href="' + '/matrix/' + this.props.query + '&award.project=' + chart.data.labels[i] + '">'); // go to matrix view when clicked
-                                text.push('<span class="chart-legend-chip" style="background-color:' + chart.data.datasets[0].backgroundColor[i] + '"></span>');
-                                if (chart.data.labels[i]) {
-                                    text.push('<span class="chart-legend-label">' + chart.data.labels[i] + '</span>');
+                    options: {
+                        responsive: true,
+                        maintainAspectRatio: false,
+                        legend: {
+                            display: false, // Hide automatically generated legend; we draw it ourselves
+                        },
+                        animation: {
+                            duration: 200,
+                        },
+                        legendCallback: (chart) => { // allows for legend clicking
+                            const chartData = chart.data.datasets[0].data;
+                            const text = [];
+                            text.push('<ul>');
+                            for (let i = 0; i < chartData.length; i += 1) {
+                                if (chartData[i]) {
+                                    text.push('<li>');
+                                    text.push(`<a href="/matrix/${this.props.query}&award.project=${chart.data.labels[i]}">`); // go to matrix view when clicked
+                                    text.push(`<span class="chart-legend-chip" style="background-color:${chart.data.datasets[0].backgroundColor[i]}"></span>`);
+                                    if (chart.data.labels[i]) {
+                                        text.push(`<span class="chart-legend-label">${chart.data.labels[i]}</span>`);
+                                    }
+                                    text.push('</a></li>');
                                 }
-                                text.push('</a></li>');
                             }
-                        }
-                        text.push('</ul>');
-                        return text.join('');
-                    }.bind(this),
-                    onClick: function (e) {
-                        // React to clicks on pie sections
-                        var activePoints = this.myPieChart.getElementAtEvent(e);
+                            text.push('</ul>');
+                            return text.join('');
+                        },
+                        onClick: (e) => {
+                            // React to clicks on pie sections
+                            const activePoints = this.myPieChart.getElementAtEvent(e);
 
-                        if (activePoints[0]) { // if click on wrong area, do nothing
-                            var clickedElementIndex = activePoints[0]._index;
-                            var term = this.myPieChart.data.labels[clickedElementIndex];
-                            this.context.navigate('/matrix/' + this.props.query + '&award.project=' + term);
-                        }
-                    }.bind(this)
-                }
-            });
+                            if (activePoints[0]) { // if click on wrong area, do nothing
+                                const clickedElementIndex = activePoints[0]._index;
+                                const term = this.myPieChart.data.labels[clickedElementIndex];
+                                this.context.navigate(`/matrix/${this.props.query}&award.project=${term}`);
+                            }
+                        },
+                    },
+                });
 
-            // Have chartjs draw the legend into the DOM.
-            document.getElementById('chart-legend').innerHTML = this.myPieChart.generateLegend();
+                // Have chartjs draw the legend into the DOM.
+                document.getElementById('chart-legend').innerHTML = this.myPieChart.generateLegend();
 
-            // Save the chart <div> height so we can set it to that value when no data's available.
-            let chartWrapperDiv = document.getElementById('chart-wrapper-1');
-            this.wrapperHeight = chartWrapperDiv.clientHeight;
-        }.bind(this));
-    },
+                // Save the chart <div> height so we can set it to that value when no data's available.
+                const chartWrapperDiv = document.getElementById('chart-wrapper-1');
+                this.wrapperHeight = chartWrapperDiv.clientHeight;
+            }
+        });
+    }
 
     // Update existing chart with new data.
-    updateChart: function (Chart, facetData) {
+    updateChart(Chart, facetData) {
         // for each item, set doc count, add to total doc count, add proper label, and assign color.
-        let colors = this.context.projectColors.colorList(facetData.map(term => term.key), { shade: 10 });
-        let data = [];
-        let labels = [];
+        const colors = this.context.projectColors.colorList(facetData.map(term => term.key), { shade: 10 });
+        const data = [];
+        const labels = [];
 
         // Convert facet data to chart data.
         facetData.forEach((term, i) => {
@@ -417,34 +480,17 @@ let HomepageChart = React.createClass({
 
         // Redraw the updated legend
         document.getElementById('chart-legend').innerHTML = Chart.generateLegend();
-    },
+    }
 
-    componentDidMount: function () {
-        // Create the chart, and assign the chart to this.myPieChart when the process finishes.
-        if (document.getElementById("myChart")) {
-            this.createChart(this.facetData);
-        }
-    },
-
-    componentDidUpdate: function () {
-        if (this.myPieChart) {
-            // Existing data updated
-            this.updateChart(this.myPieChart, this.facetData);
-        } else if (this.facetData.length) {
-            // Chart existed but was destroyed for lack of data. Rebuild the chart.
-            this.createChart(this.facetData);
-        }
-    },
-
-    render: function () {
-        let facets = this.props.data && this.props.data.facets;
+    render() {
+        const facets = this.props.data && this.props.data.facets;
         let total;
 
         // Get all project facets, or an empty array if none.
         if (facets) {
-            let projectFacet = facets.find(facet => facet.field === 'award.project');
+            const projectFacet = facets.find(facet => facet.field === 'award.project');
             this.facetData = projectFacet ? projectFacet.terms : [];
-            let docCounts = this.facetData.length ? this.facetData.map(data => data.doc_count) : [];
+            const docCounts = this.facetData.length ? this.facetData.map(data => data.doc_count) : [];
             total = docCounts.length ? docCounts.reduce((prev, curr) => prev + curr) : 0;
 
             // No data with the current selection, but we used to? Destroy the existing chart so we can
@@ -465,14 +511,14 @@ let HomepageChart = React.createClass({
             <div>
                 <div className="title">
                     Project
-                    <center> <hr width="80%" position="static" color="blue"></hr> </center>
+                    <center><hr width="80%" color="blue" /></center>
                 </div>
                 {this.facetData.length && total ?
                     <div id="chart-wrapper-1" className="chart-wrapper">
                         <div className="chart-container">
-                            <canvas id="myChart"></canvas>
+                            <canvas id="myChart" />
                         </div>
-                        <div id="chart-legend" className="chart-legend"></div>
+                        <div id="chart-legend" className="chart-legend" />
                     </div>
                     :
                     <div className="chart-no-data" style={{ height: this.wrapperHeight }}><p>No data to display</p></div>
@@ -480,31 +526,54 @@ let HomepageChart = React.createClass({
             </div>
         );
     }
+}
 
-});
+HomepageChart.propTypes = {
+    query: PropTypes.string,
+    data: PropTypes.object,
+};
+
+HomepageChart.contextTypes = {
+    navigate: PropTypes.func,
+    projectColors: PropTypes.object, // DataColor instance for experiment project
+};
 
 
 // Component to display the D3-based chart for Biosample
-var HomepageChart2 = React.createClass({
+class HomepageChart2 extends React.Component {
+    constructor(props) {
+        super(props);
+        this.wrapperHeight = 200;
+        this.createChart = this.createChart.bind(this);
+        this.updateChart = this.updateChart.bind(this);
+    }
 
-    contextTypes: {
-        navigate: React.PropTypes.func,
-        biosampleTypeColors: React.PropTypes.object // DataColor instance for experiment project
-    },
+    componentDidMount() {
+        if (document.getElementById('myChart2')) {
+            this.createChart(this.facetData);
+        }
+    }
 
-    wrapperHeight: 200,
+    componentDidUpdate() {
+        if (this.myPieChart) {
+            // Existing data updated
+            this.updateChart(this.myPieChart, this.facetData);
+        } else if (this.facetData.length) {
+            // Chart existed but was destroyed for lack of data. Rebuild the chart.
+            this.createChart(this.facetData);
+        }
+    }
 
-    createChart: function (facetData) {
-
+    createChart(facetData) {
         // Draw the chart of search results given in this.props.data.facets. Since D3 doesn't work
         // with the React virtual DOM, we have to load it separately using the webpack .ensure
         // mechanism. Once the callback is called, it's loaded and can be referenced through
         // require.
-        require.ensure(['chart.js'], function (require) {
-            var Chart = require('chart.js');
-            var colors = this.context.biosampleTypeColors.colorList(facetData.map(term => term.key), { shade: 10 });
-            var data = [];
-            var labels = [];
+        require.ensure(['chart.js'], (require) => {
+            const Chart = require('chart.js');
+            const colors = this.context.biosampleTypeColors.colorList(facetData.map(term => term.key), { shade: 10 });
+            const data = [];
+            const labels = [];
 
             facetData.forEach((term, i) => {
                 data[i] = term.doc_count;
@@ -514,37 +583,37 @@ var HomepageChart2 = React.createClass({
             // adding total doc count to middle of donut
             // http://stackoverflow.com/questions/20966817/how-to-add-text-inside-the-doughnut-chart-using-chart-js/24671908
             Chart.pluginService.register({
-                beforeDraw: drawDonutCenter
+                beforeDraw: drawDonutCenter,
             });
 
             // Pass the assay_title counts to the charting library to render it.
-            var canvas = document.getElementById("myChart2");
+            const canvas = document.getElementById('myChart2');
             if (canvas) {
-                var ctx = canvas.getContext("2d");
+                const ctx = canvas.getContext('2d');
                 this.myPieChart = new Chart(ctx, {
                     type: 'doughnut',
                     data: {
                         labels: labels,
                         datasets: [{
                             data: data,
-                            backgroundColor: colors
-                        }]
+                            backgroundColor: colors,
+                        }],
                     },
                     options: {
                         responsive: true,
                         maintainAspectRatio: false,
                         legend: {
-                            display: false // hiding automatically generated legend
+                            display: false, // hiding automatically generated legend
                         },
                         animation: {
-                            duration: 200
+                            duration: 200,
                         },
-                        legendCallback: function (chart) { // allows for legend clicking
-                            let data = chart.data.datasets[0].data;
-                            let text = [];
+                        legendCallback: (chart) => { // allows for legend clicking
+                            const chartData = chart.data.datasets[0].data;
+                            const text = [];
                             text.push('<ul>');
-                            for (let i = 0; i < data.length; i++) {
-                                if (data[i]) {
+                            for (let i = 0; i < chartData.length; i += 1) {
+                                if (chartData[i]) {
                                     text.push('<li>');
                                     text.push(`<a href="/matrix/${this.props.query}&biosample_type=${chart.data.labels[i]}">`); // go to matrix view when clicked
                                     text.push(`<span class="chart-legend-chip" style="background-color:${chart.data.datasets[0].backgroundColor[i]}"></span>`);
@@ -556,17 +625,18 @@ var HomepageChart2 = React.createClass({
                             }
                             text.push('</ul>');
                             return text.join('');
-                        }.bind(this),
-                        onClick: function (e) {
+                        },
+                        onClick: (e) => {
                             // React to clicks on pie sections
-                            let activePoints = this.myPieChart.getElementAtEvent(e);
+                            const query = this.computationalPredictions ? 'biosample_type=' : 'replicates.library.biosample.biosample_type=';
+                            const activePoints = this.myPieChart.getElementAtEvent(e);
                             if (activePoints[0]) {
-                                let clickedElementIndex = activePoints[0]._index;
-                                let term = this.myPieChart.data.labels[clickedElementIndex];
-                                this.context.navigate(`/matrix/${this.props.query}&biosample_type=${term}`); // go to matrix view
+                                const clickedElementIndex = activePoints[0]._index;
+                                const term = this.myPieChart.data.labels[clickedElementIndex];
+                                this.context.navigate(`/matrix/${this.props.query}&${query}${term}`); // go to matrix view
                             }
-                        }.bind(this)
-                    }
+                        },
+                    },
                 });
             } else {
                 this.myPieChart = null;
@@ -579,18 +649,18 @@ var HomepageChart2 = React.createClass({
             }
 
             // Save the chart <div> height so we can set it to that value when no data's available.
-            let chartWrapperDiv = document.getElementById('chart-wrapper-2');
+            const chartWrapperDiv = document.getElementById('chart-wrapper-2');
             if (chartWrapperDiv) {
                 this.wrapperHeight = chartWrapperDiv.clientHeight;
             }
-        }.bind(this));
-    },
+        });
+    }
 
-    updateChart: function (Chart, facetData) {
+    updateChart(Chart, facetData) {
         // for each item, set doc count, add to total doc count, add proper label, and assign color.
-        let colors = this.context.biosampleTypeColors.colorList(facetData.map(term => term.key), { shade: 10 });
-        let data = [];
-        let labels = [];
+        const colors = this.context.biosampleTypeColors.colorList(facetData.map(term => term.key), { shade: 10 });
+        const data = [];
+        const labels = [];
 
         // Convert facet data to chart data.
         facetData.forEach((term, i) => {
@@ -606,35 +676,18 @@ var HomepageChart2 = React.createClass({
 
         // Redraw the updated legend
         document.getElementById('chart-legend-2').innerHTML = Chart.generateLegend(); // generates legend
-    },
+    }
 
-    componentDidMount: function () {
-        if (document.getElementById("myChart2")) {
-            this.createChart(this.facetData);
-        }
-    },
-
-    componentDidUpdate: function () {
-        if (this.myPieChart) {
-            // Existing data updated
-            this.updateChart(this.myPieChart, this.facetData);
-        } else if (this.facetData.length) {
-            // Chart existed but was destroyed for lack of data. Rebuild the chart.
-            this.createChart(this.facetData);
-        }
-    },
-
-    render: function () {
-        let assayFacet = {};
-        let facets = this.props.data && this.props.data.facets;
+    render() {
+        const facets = this.props.data && this.props.data.facets;
         let total;
 
         // Our data source will be different for computational predictions
         if (facets) {
             this.computationalPredictions = this.props.assayCategory === 'COMPPRED';
-            assayFacet = facets.find(facet => facet.field === 'biosample_type');
+            const assayFacet = facets.find(facet => facet.field === 'biosample_type');
             this.facetData = assayFacet ? assayFacet.terms : [];
-            let docCounts = this.facetData.length ? this.facetData.map(data => data.doc_count) : [];
+            const docCounts = this.facetData.length ? this.facetData.map(data => data.doc_count) : [];
             total = docCounts.length ? docCounts.reduce((prev, curr) => prev + curr) : 0;
 
             // No data with the current selection, but we used to destroy the existing chart so we can
@@ -655,14 +708,14 @@ var HomepageChart2 = React.createClass({
             <div>
                 <div className="title">
                     Biosample Type
-                    <center> <hr width="80%" position="static" color="blue"></hr> </center>
+                    <center><hr width="80%" color="blue" /></center>
                 </div>
                 {this.facetData.length && total ?
                     <div id="chart-wrapper-2" className="chart-wrapper">
                         <div className="chart-container">
-                            <canvas id="myChart2"></canvas>
+                            <canvas id="myChart2" />
                         </div>
-                        <div id="chart-legend-2" className="chart-legend"></div>
+                        <div id="chart-legend-2" className="chart-legend" />
                     </div>
                     :
                     <div className="chart-no-data" style={{ height: this.wrapperHeight }}>No data to display</div>
@@ -670,7 +723,18 @@ var HomepageChart2 = React.createClass({
             </div>
         );
     }
-});
+}
+
+HomepageChart2.propTypes = {
+    query: PropTypes.string,
+    data: PropTypes.object,
+    assayCategory: PropTypes.string,
+};
+
+HomepageChart2.contextTypes = {
+    navigate: PropTypes.func,
+    biosampleTypeColors: PropTypes.object, // DataColor instance for experiment project
+};
 
 
 // Draw the small triangle above the selected assay in the "Assay Categories" chart if the user has
@@ -700,25 +764,40 @@ function drawColumnSelects(currentAssay, ctx, data) {
 
 
 // Component to display the D3-based chart for Biosample
-let HomepageChart3 = React.createClass({
+class HomepageChart3 extends React.Component {
+    constructor(props) {
+        super(props);
+        this.wrapperHeight = 200;
+        this.createChart = this.createChart.bind(this);
+        this.updateChart = this.updateChart.bind(this);
+    }
 
-    contextTypes: {
-        navigate: React.PropTypes.func
-    },
+    componentDidMount() {
+        if (document.getElementById('myChart3')) {
+            this.createChart(this.facetData);
+        }
+    }
 
-    wrapperHeight: 200,
+    componentDidUpdate() {
+        if (this.myPieChart) {
+            // Existing data updated
+            this.updateChart(this.myPieChart, this.facetData);
+        } else if (this.facetData.length) {
+            // Chart existed but was destroyed for lack of data. Rebuild the chart.
+            this.createChart(this.facetData);
+        }
+    }
 
-    createChart: function (facetData) {
-
+    createChart(facetData) {
         // Draw the chart of search results given in this.props.data.facets. Since D3 doesn't work
         // with the React virtual DOM, we have to load it separately using the webpack .ensure
         // mechanism. Once the callback is called, it's loaded and can be referenced through
         // require.
-        require.ensure(['chart.js'], function (require) {
-            let Chart = require('chart.js');
-            let colors = [];
-            let data = [];
-            let labels = [];
+        require.ensure(['chart.js'], (require) => {
+            const Chart = require('chart.js');
+            const colors = [];
+            const data = [];
+            const labels = [];
             const selectedAssay = (this.props.assayCategory && this.props.assayCategory !== 'COMPPRED') ? this.props.assayCategory.replace(/\+/g, ' ') : '';
 
             // For each item, set doc count, add to total doc count, add proper label, and assign
@@ -730,74 +809,75 @@ let HomepageChart3 = React.createClass({
             });
 
             // Pass the counts to the charting library to render it.
-            let canvas = document.getElementById("myChart3");
-            let ctx = canvas.getContext("2d");
-            this.myPieChart = new Chart(ctx, {
-                type: 'bar',
-                data: {
-                    labels: labels, // full labels
-                    datasets: [{
-                        data: data,
-                        backgroundColor: colors
-                    }]
-                },
-                options: {
-                    responsive: true,
-                    maintainAspectRatio: false,
-                    legend: {
-                        display: false // hiding automatically generated legend
-                    },
-                    hover: {
-                        mode: false,
-                    },
-                    animation: {
-                        duration: 0,
-                        onProgress: function() { drawColumnSelects(selectedAssay, this.chart.ctx, this.data); },
-                        onComplete: function() { drawColumnSelects(selectedAssay, this.chart.ctx, this.data); },
-                    },
-                    scales: {
-                        xAxes: [{
-                            gridLines: {
-                                display: false
-                            },
-                            ticks: {
-                                autoSkip: false
-                            }
+            const canvas = document.getElementById('myChart3');
+            if (canvas) {
+                const ctx = canvas.getContext('2d');
+                this.myPieChart = new Chart(ctx, {
+                    type: 'bar',
+                    data: {
+                        labels: labels, // full labels
+                        datasets: [{
+                            data: data,
+                            backgroundColor: colors,
                         }],
                     },
-                    layout: {
-                        padding: {
-                            top: 10,
-                        }
+                    options: {
+                        responsive: true,
+                        maintainAspectRatio: false,
+                        legend: {
+                            display: false, // hiding automatically generated legend
+                        },
+                        hover: {
+                            mode: false,
+                        },
+                        animation: {
+                            duration: 0,
+                            onProgress: function () { drawColumnSelects(selectedAssay, this.chart.ctx, this.data); },
+                            onComplete: function () { drawColumnSelects(selectedAssay, this.chart.ctx, this.data); },
+                        },
+                        scales: {
+                            xAxes: [{
+                                gridLines: {
+                                    display: false,
+                                },
+                                ticks: {
+                                    autoSkip: false,
+                                },
+                            }],
+                        },
+                        layout: {
+                            padding: {
+                                top: 10,
+                            },
+                        },
+                        onClick: (e) => {
+                            // React to clicks on pie sections
+                            const query = 'assay_slims=';
+                            const activePoints = this.myPieChart.getElementAtEvent(e);
+                            if (activePoints[0]) {
+                                const clickedElementIndex = activePoints[0]._index;
+                                const term = this.myPieChart.data.labels[clickedElementIndex];
+                                this.context.navigate(`/matrix/${this.props.query}&${query}${term}`); // go to matrix view
+                            }
+                        },
                     },
-                    onClick: function (e) {
-                        // React to clicks on pie sections
-                        var query = 'assay_slims=';
-                        var activePoints = this.myPieChart.getElementAtEvent(e);
-                        if (activePoints[0]) {
-                            let clickedElementIndex = activePoints[0]._index;
-                            var term = this.myPieChart.data.labels[clickedElementIndex];
-                            this.context.navigate('/matrix/' + this.props.query + '&' + query + term); // go to matrix view
-                        }
-                    }.bind(this)
-                }
-            });
+                });
 
-            // Save height of wrapper div.
-            let chartWrapperDiv = document.getElementById('chart-wrapper-3');
-            this.wrapperHeight = chartWrapperDiv.clientHeight;
-        }.bind(this));
+                // Save height of wrapper div.
+                const chartWrapperDiv = document.getElementById('chart-wrapper-3');
+                this.wrapperHeight = chartWrapperDiv.clientHeight;
+            }
+        });
+    }
 
-    },
-
-    updateChart: function (Chart, facetData) {
+    updateChart(Chart, facetData) {
         // for each item, set doc count, add to total doc count, add proper label, and assign color.
-        let data = [];
-        let labels = [];
-        let colors = [];
+        const data = [];
+        const labels = [];
+        const colors = [];
 
         // Convert facet data to chart data.
-        let selectedAssay = (this.props.assayCategory && this.props.assayCategory !== 'COMPPRED') ? this.props.assayCategory.replace(/\+/g, ' ') : '';
+        const selectedAssay = (this.props.assayCategory && this.props.assayCategory !== 'COMPPRED') ? this.props.assayCategory.replace(/\+/g, ' ') : '';
         facetData.forEach((term, i) => {
             data[i] = term.doc_count;
             labels[i] = term.key;
@@ -809,36 +889,20 @@ let HomepageChart3 = React.createClass({
         Chart.data.labels = labels;
         Chart.data.datasets[0].backgroundColor = colors;
         Chart.options.hover.mode = false;
-        Chart.options.animation.onProgress = function() { drawColumnSelects(selectedAssay, this.chart.ctx, this.data); }
-        Chart.options.animation.onComplete = function() { drawColumnSelects(selectedAssay, this.chart.ctx, this.data); }
+        Chart.options.animation.onProgress = function () { drawColumnSelects(selectedAssay, this.chart.ctx, this.data); };
+        Chart.options.animation.onComplete = function () { drawColumnSelects(selectedAssay, this.chart.ctx, this.data); };
         Chart.update();
-    },
+    }
 
-    componentDidMount: function () {
-        if (document.getElementById("myChart3")) {
-            this.createChart(this.facetData);
-        }
-    },
-
-    componentDidUpdate: function () {
-        if (this.myPieChart) {
-            // Existing data updated
-            this.updateChart(this.myPieChart, this.facetData);
-        } else if (this.facetData.length) {
-            // Chart existed but was destroyed for lack of data. Rebuild the chart.
-            this.createChart(this.facetData);
-        }
-    },
-
-    render: function () {
-        let facets = this.props.data && this.props.data.facets;
+    render() {
+        const facets = this.props.data && this.props.data.facets;
         let total;
 
         // Get all assay category facets, or an empty array if none
         if (facets) {
-            let projectFacet = facets.find(facet => facet.field === 'assay_slims');
+            const projectFacet = facets.find(facet => facet.field === 'assay_slims');
             this.facetData = projectFacet ? projectFacet.terms : [];
-            let docCounts = this.facetData.length ? this.facetData.map(data => data.doc_count) : [];
+            const docCounts = this.facetData.length ? this.facetData.map(data => data.doc_count) : [];
             total = docCounts.length ? docCounts.reduce((prev, curr) => prev + curr) : 0;
 
             // No data with the current selection, but we used to? Destroy the existing chart so we can
@@ -859,12 +923,12 @@ let HomepageChart3 = React.createClass({
             <div>
                 <div className="title">
                     Assay Categories
-                    <center> <hr width="80%" position="static" color="blue"></hr> </center>
+                    <center><hr width="80%" color="blue" /></center>
                 </div>
                 {this.facetData.length && total ?
                     <div id="chart-wrapper-3" className="chart-wrapper">
                         <div className="chart-container-assaycat">
-                            <canvas id="myChart3"></canvas>
+                            <canvas id="myChart3" />
                         </div>
                     </div>
                     :
@@ -873,98 +937,121 @@ let HomepageChart3 = React.createClass({
             </div>
         );
     }
+}
 
-});
+HomepageChart3.propTypes = {
+    assayCategory: PropTypes.string,
+    query: PropTypes.string,
+    data: PropTypes.object,
+};
 
-
-// Send a GET request for the most recent five news posts.
-var NewsLoader = React.createClass({
-    propTypes: {
-        newsLoaded: React.PropTypes.func.isRequired // Called parent once the news is loaded
-    },
-
-    render: function () {
-        return <FetchedItems {...this.props} url={newsUri + '&limit=5'} Component={News} ignoreErrors newsLoaded={this.props.newsLoaded} />;
-    }
-});
+HomepageChart3.contextTypes = {
+    navigate: PropTypes.func,
+};
 
 
 // Render the most recent five news posts
-var News = React.createClass({
-    propTypes: {
-        newsLoaded: React.PropTypes.func.isRequired // Called parent once the news is loaded
-    },
+class News extends React.Component {
+    componentDidMount() {
+        this.props.newsLoaded(this.nodeRef);
+    }
 
-    componentDidMount: function () {
-        this.props.newsLoaded();
-    },
-
-    render: function () {
-        var items = this.props.items;
+    render() {
+        const { items, nodeRef } = this.props;
         if (items && items.length) {
             return (
-                <div className="news-listing">
-                    <NewsPreviews items={items} />
+                <div ref={(node) => { this.nodeRef = node; }} className="news-listing">
+                    {items.map(item =>
+                        <div key={item['@id']} className="news-listing-item">
+                            <h3>{item.title}</h3>
+                            <h4>{moment.utc(item.date_created).format('MMMM D, YYYY')}</h4>
+                            <div className="news-excerpt">{item.news_excerpt}</div>
+                            <div className="news-listing-readmore">
+                                <a className="btn btn-info btn-sm" href={item['@id']} title={`View news post for ${item.title}`} key={item['@id']}>Read more</a>
+                            </div>
+                        </div>,
+                    )}
                 </div>
             );
-        } else {
-            return <div className="news-empty">No news available at this time</div>
+        }
+        return <div className="news-empty">No news available at this time</div>;
+    }
+}
+
+News.propTypes = {
+    items: PropTypes.array,
+    newsLoaded: PropTypes.func.isRequired, // Called parent once the news is loaded
+    nodeRef: PropTypes.func, // React ref callback so we can get the news-listing DOM element in a higher component
+};
+
+
+// Send a GET request for the most recent five news posts. Don't make this a stateless component
+// because we attach `ref` to this, and stateless components don't support that.
+class NewsLoader extends React.Component {
+    render() {
+        return <FetchedItems {...this.props} url={`${newsUri}&limit=5`} Component={News} ignoreErrors newsLoaded={this.props.newsLoaded} />;
+    }
+}
+
+NewsLoader.propTypes = {
+    newsLoaded: PropTypes.func.isRequired, // Called parent once the news is loaded
+};
+
+
+class TwitterWidget extends React.Component {
+    constructor(props) {
+        super(props);
+        this.initialized = false;
+        this.injectTwitter = this.injectTwitter.bind(this);
+    }
+
+    componentDidMount() {
+        if (!this.initialized && this.props.height) {
+            this.injectTwitter();
         }
     }
-});
 
+    componentDidUpdate() {
+        if (!this.initialized && this.props.height) {
+            this.injectTwitter();
+        }
+    }
 
-var TwitterWidget = React.createClass({
-    propTypes: {
-        height: React.PropTypes.number.isRequired // Number of pixels tall to make widget
-    },
-
-    injectTwitter: function () {
-        var js, link;
+    injectTwitter() {
         if (!this.initialized) {
-            link = this.refs.link.getDOMNode();
+            const link = this.anchor;
             this.initialized = true;
-            js = document.createElement("script");
-            js.id = "twitter-wjs";
-            js.src = "//platform.twitter.com/widgets.js";
+            const js = document.createElement('script');
+            js.id = 'twitter-wjs';
+            js.src = '//platform.twitter.com/widgets.js';
             return link.parentNode.appendChild(js);
         }
-    },
+        return null;
+    }
 
-    componentDidMount: function () { // twitter script from API
-        if (!this.initialized && this.props.height) {
-            this.injectTwitter();
-        }
-    },
-
-    componentDidUpdate: function () {
-        if (!this.initialized && this.props.height) {
-            this.injectTwitter();
-        }
-    },
-
-    render: function () {
-        var content, ref2, title, widget;
+    render() {
         return (
             <div ref="twitterwidget">
                 <div className="twitter-header">
-                    <h2>Twitter <a href="https://twitter.com/EncodeDCC" title="ENCODE DCC Twitter page in a new window or tab" target="_blank" className="twitter-ref">@EncodeDCC</a></h2>
+                    <h2>Twitter <a href="https://twitter.com/EncodeDCC" title="ENCODE DCC Twitter page in a new window or tab" target="_blank" rel="noopener noreferrer"className="twitter-ref">@EncodeDCC</a></h2>
                 </div>
                 {this.props.height ?
                     <a
-                        ref="link"
+                        ref={(anchor) => { this.anchor = anchor; }}
                         className="twitter-timeline"
                         href="https://twitter.com/encodedcc" // from encodedcc twitter
-                        widget-id="encodedcc"
                         data-chrome="noheader"
                         data-screen-name="EncodeDCC"
-                        //data-tweet-limit = "4"
-                        //data-width = "300"
-                        data-height={this.props.height + ''} // height so it matches with rest of site
-                        ></a>
-                    : null}
+                        data-height={this.props.height.toString()} // height so it matches with rest of site
+                    >
+                        @EncodeDCC
+                    </a>
+                : null}
             </div>
         );
     }
-});
+}
 
+TwitterWidget.propTypes = {
+    height: PropTypes.number.isRequired, // Number of pixels tall to make widget
+};
