@@ -1,4 +1,6 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import createReactClass from 'create-react-class';
 import moment from 'moment';
 import querystring from 'querystring';
 import _ from 'underscore';
@@ -8,9 +10,9 @@ import { Panel, PanelBody } from '../libs/bootstrap/panel';
 
 
 // Display a news preview item from a search result @graph object.
-const NewsPreviewItem = React.createClass({
+const NewsPreviewItem = createReactClass({
     propTypes: {
-        item: React.PropTypes.object, // News search result object to display
+        item: PropTypes.object, // News search result object to display
     },
 
     render: function () {
@@ -30,9 +32,9 @@ const NewsPreviewItem = React.createClass({
 
 
 // Display a list of news preview items from search results.
-const NewsPreviews = React.createClass({
+const NewsPreviews = createReactClass({
     propTypes: {
-        items: React.PropTypes.array, // Items from search result @graph
+        items: PropTypes.array, // Items from search result @graph
     },
 
     render: function () {
@@ -54,11 +56,11 @@ export default NewsPreviews;
 
 // Default news facet-rendering component for any type of facet not specifically registered to render
 // in its own component, so the vast majority of news facets use this component.
-const DefaultFacet = React.createClass({
+const DefaultFacet = createReactClass({
     propTypes: {
-        facet: React.PropTypes.object.isRequired, // One facet from search results
-        baseUri: React.PropTypes.string.isRequired, // Current URI and query string
-        searchFilters: React.PropTypes.object.isRequired, // Search-result filter object; see <NewsFacets />
+        facet: PropTypes.object.isRequired, // One facet from search results
+        baseUri: PropTypes.string.isRequired, // Current URI and query string
+        searchFilters: PropTypes.object.isRequired, // Search-result filter object; see <NewsFacets />
     },
 
     render: function () {
@@ -112,11 +114,11 @@ globals.facet_view.fallback = () => DefaultFacet;
 
 
 // Special news facet-rendering component to render facets with facet.field==='month_released'.
-const DateFacet = React.createClass({
+const DateFacet = createReactClass({
     propTypes: {
-        facet: React.PropTypes.object.isRequired, // One facet from search results
-        baseUri: React.PropTypes.string.isRequired, // Current URI and query string
-        searchFilters: React.PropTypes.object.isRequired, // Search-result filter object; see <NewsFacets />
+        facet: PropTypes.object.isRequired, // One facet from search results
+        baseUri: PropTypes.string.isRequired, // Current URI and query string
+        searchFilters: PropTypes.object.isRequired, // Search-result filter object; see <NewsFacets />
     },
 
     render: function () {
@@ -301,11 +303,11 @@ globals.facet_view.register(DateFacet, 'month_released');
 
 
 // Display the news facets.
-const NewsFacets = React.createClass({
+const NewsFacets = createReactClass({
     propTypes: {
-        facets: React.PropTypes.array.isRequired, // Array of facets, each containing an array of facet items
-        filters: React.PropTypes.array.isRequired, // Array of filters from the search results
-        baseUri: React.PropTypes.string.isRequired, // Current URI and query string
+        facets: PropTypes.array.isRequired, // Array of facets, each containing an array of facet items
+        filters: PropTypes.array.isRequired, // Array of filters from the search results
+        baseUri: PropTypes.string.isRequired, // Current URI and query string
     },
 
     render: function () {
@@ -355,9 +357,9 @@ const NewsFacets = React.createClass({
 
 
 // Rendering component for the stand-alone news preview page.
-const News = React.createClass({
+const News = createReactClass({
     propTypes: {
-        context: React.PropTypes.object, // Search results
+        context: PropTypes.object, // Search results
     },
 
     render: function () {
