@@ -59,6 +59,8 @@
 // }
 
 var React = require('react');
+import PropTypes from 'prop-types';
+import createReactClass from 'create-react-class';
 var _ = require('underscore');
 var moment = require('moment');
 var panel = require('../libs/bootstrap/panel');
@@ -68,15 +70,15 @@ var {Panel, PanelHeading} = panel;
 
 // Required sortable table wrapper component. Takes no parameters but puts the table in a Bootstrap panel
 // and makes it responsive. You can place multiple <SortTable />s as children of this component.
-var SortTablePanel = module.exports.SortTablePanel = React.createClass({
+var SortTablePanel = module.exports.SortTablePanel = createReactClass({
     propTypes: {
         // Note: `title` overrides `header`
-        title: React.PropTypes.oneOfType([ // Title to display in table panel header
-            React.PropTypes.string, // When title is a simple string
-            React.PropTypes.object // When title is JSX
+        title: PropTypes.oneOfType([ // Title to display in table panel header
+            PropTypes.string, // When title is a simple string
+            PropTypes.object // When title is JSX
         ]),
-        header: React.PropTypes.object, // React component to render inside header
-        noDefaultClasses: React.PropTypes.bool // T to skip default <Panel> classes
+        header: PropTypes.object, // React component to render inside header
+        noDefaultClasses: PropTypes.bool // T to skip default <Panel> classes
     },
 
     render: function() {
@@ -101,14 +103,14 @@ var SortTablePanel = module.exports.SortTablePanel = React.createClass({
 });
 
 
-var SortTableComponent = module.exports.SortTableComponent = React.createClass({
+var SortTableComponent = module.exports.SortTableComponent = createReactClass({
     propTypes: {
         // Note: `title` overrides `header`
-        title: React.PropTypes.oneOfType([ // Title to display in table panel header
-            React.PropTypes.string, // When title is a simple string
-            React.PropTypes.object // When title is JSX
+        title: PropTypes.oneOfType([ // Title to display in table panel header
+            PropTypes.string, // When title is a simple string
+            PropTypes.object // When title is JSX
         ]),
-        header: React.PropTypes.object // React component to render inside header
+        header: PropTypes.object // React component to render inside header
     },
 
     render: function() {
@@ -132,18 +134,18 @@ var SortTableComponent = module.exports.SortTableComponent = React.createClass({
 
 
 // Displays one table within a <SortTablePanel></SortTablePanel>.
-var SortTable = module.exports.SortTable = React.createClass({
+var SortTable = module.exports.SortTable = createReactClass({
     propTypes: {
-        title: React.PropTypes.oneOfType([ // Title to display in table header
-            React.PropTypes.string, // When title is a simple string
-            React.PropTypes.object // When title is JSX
+        title: PropTypes.oneOfType([ // Title to display in table header
+            PropTypes.string, // When title is a simple string
+            PropTypes.object // When title is JSX
         ]),
-        list: React.PropTypes.array, // Array of objects to display in the table
-        columns: React.PropTypes.object.isRequired, // Defines the columns of the table
-        rowClasses: React.PropTypes.func, // If provided, gets called for each row of table to generate per-row CSS classes
-        sortColumn: React.PropTypes.string, // ID of column to sort by default; first column if not given
-        footer: React.PropTypes.object, // Optional component to display in the footer
-        collapsed: React.PropTypes.bool // T if only title bar should be displayed
+        list: PropTypes.array, // Array of objects to display in the table
+        columns: PropTypes.object.isRequired, // Defines the columns of the table
+        rowClasses: PropTypes.func, // If provided, gets called for each row of table to generate per-row CSS classes
+        sortColumn: PropTypes.string, // ID of column to sort by default; first column if not given
+        footer: PropTypes.object, // Optional component to display in the footer
+        collapsed: PropTypes.bool // T if only title bar should be displayed
     },
 
     getInitialState: function() {
