@@ -1,7 +1,10 @@
 'use strict';
 var React = require('react');
+import PropTypes from 'prop-types';
+import createReactClass from 'create-react-class';
 var panel = require('../libs/bootstrap/panel');
 var globals = require('./globals');
+var { PickerActions } = require('./search');
 var search = require('./search');
 import StatusLabel from './statuslabel';
 var navigation = require('./navigation');
@@ -25,7 +28,7 @@ var PanelLookup = function (props) {
 
 
 // Display a stand-alone TALEN page.
-var TalenPage = React.createClass({
+var TalenPage = createReactClass({
     render: function() {
         var context = this.props.context;
         var talen = PanelLookup({context: context});
@@ -66,9 +69,9 @@ globals.content_views.register(TalenPage, 'TALEN');
 
 
 // Display a TALENs panel. We can have multiple TALENs in one panel
-var Talen = React.createClass({
+var Talen = createReactClass({
     propTypes: {
-        context: React.PropTypes.object.isRequired // TALEN object
+        context: PropTypes.object.isRequired // TALEN object
     },
 
     render: function() {
@@ -122,9 +125,7 @@ globals.panel_views.register(Talen, 'TALEN');
 
 
 // Search result page output
-var Listing = React.createClass({
-    mixins: [search.PickerActionsMixin],
-
+var Listing = createReactClass({
     render: function() {
         var result = this.props.context;
         var coordinates = result.target_genomic_coordinates;
@@ -132,7 +133,7 @@ var Listing = React.createClass({
         return (
             <li>
                 <div className="clearfix">
-                    {this.renderActions()}
+                    <PickerActions {...this.props} />
                     <div className="pull-right search-meta">
                         <p className="type meta-title">TALEN</p>
                         {result.status ? <p className="type meta-status">{' ' + result.status}</p> : ''}
