@@ -1,5 +1,7 @@
 'use strict';
 var React = require('react');
+import PropTypes from 'prop-types';
+import createReactClass from 'create-react-class';
 
 
 // Render a dropdown menu. All components within the dropdown get wrapped in <li> tags, so the 'a'
@@ -18,9 +20,11 @@ var React = require('react');
 // <li> --separator-- </li> (well, actually a line, not the word "separator")
 // <li><a href="#">Third</a></li>
 
-module.exports.DropdownMenu = React.createClass({
+module.exports.DropdownMenu = createReactClass({
+    // One might think `label` should be isRequired. But we can't because of:
+    // https://github.com/facebook/react/issues/4494#issuecomment-125068868
     propTypes: {
-        label: React.PropTypes.string.isRequired // id attribute value for the button that controls this menu
+        label: PropTypes.string, // id attribute value for the button that controls this menu
     },
 
     render: function() {
@@ -35,8 +39,8 @@ module.exports.DropdownMenu = React.createClass({
 });
 
 
-module.exports.DropdownMenuSep = React.createClass({
+module.exports.DropdownMenuSep = createReactClass({
     render: function() {
-        return <li className="dropdown-sep"></li>;
+        return <div className="dropdown-sep"></div>;
     }
 });
