@@ -1,20 +1,14 @@
-'use strict';
+import React from 'react';
+import { mount } from 'enzyme';
 
-jest.autoMockOff();
+// Import test component and data.
+import { Lot as Antibody } from '../antibody';
+import context from '../testdata/antibody/ENCAB000AUZ';
 
-// Fixes https://github.com/facebook/jest/issues/78
-jest.dontMock('react');
 
-describe('Antibody', function() {
-    var React, TestUtils, Antibody, context;
-
-    beforeEach(function() {
-        React = require('react');
-        Antibody = require('../antibody').Lot;
-        TestUtils = require('react/lib/ReactTestUtils');
-
+describe('Antibody', () => {
+    beforeAll(() => {
         // Set up context object to be rendered
-        context = require('../testdata/antibody/ENCAB000AUZ');
         context.host_organism = require('../testdata/organism/rabbit');
         context.source = require('../testdata/source/aviva');
         context.characterizations = [require('../testdata/characterization/antibody-367a6fdd0cef')];
@@ -30,84 +24,73 @@ describe('Antibody', function() {
         context.lot_reviews[0].targets[0].organisms = [require('../testdata/organism/human')];
     });
 
-    describe('Typical antibody', function() {
-        var antibody;
+    describe('Typical antibody', () => {
+        let antibody;
 
-        beforeEach(function() {
-            var Context = {
-              fetch: function(url, options) {
-                return Promise.resolve({json: () => ({'@graph': []})});
-              }
-            };
-
+        beforeAll(() => {
             // Render antibody into jsnode
-            antibody = React.withContext(Context, function() {
-                return TestUtils.renderIntoDocument(
-                    <Antibody context={context} />
-                );
-            });
+            antibody = mount(
+                <Antibody context={context} />
+            );
         });
 
-        it('has a good header', function() {
-            var headerLine = TestUtils.findRenderedDOMComponentWithTag(antibody, 'h2').getDOMNode();
-            expect(headerLine.textContent).toEqual('ENCAB000AUZ');
-            headerLine = headerLine.nextSibling;
-            expect(headerLine.textContent).toEqual('Antibody against Homo sapiens HNRNPA1');
+        test('has a good header', () => {
+            let headerLine = antibody.find('h2');
+            expect(headerLine.text()).toEqual('ENCAB000AUZ');
+            headerLine = antibody.find('h3');
+            expect(headerLine.text()).toEqual('Antibody against Homo sapiens HNRNPA1');
         });
 
-        it('has a good status panel', function() {
-            var panel = TestUtils.findRenderedDOMComponentWithClass(antibody, 'type-antibody-status').getDOMNode();
-            var row = panel.getElementsByClassName('status-organism-row');
-            var element = row[0].getElementsByClassName('status-awaiting-lab-characterization');
-            expect(element.length).toEqual(1);
-            element = element[0].nextSibling;
-            expect(element.textContent).toEqual('awaiting lab characterization');
-            element = row[0].getElementsByClassName('status-organism');
-            expect(element[0].textContent).toEqual('Homo sapiens');
-            element = row[0].getElementsByClassName('status-terms');
-            expect(element[0].textContent).toEqual('HeLa-S3');
+        test('has a good status panel', () => {
+            const panel = antibody.find('.type-antibody-status');
+            const row = panel.find('.status-organism-row');
+            const element = row.at(0).find('.status-status');
+            expect(element).toHaveLength(1);
+            expect(element.text()).toEqual('awaiting lab characterization');
+            expect(row.find('.status-organism').text()).toEqual('Homo sapiens');
+            expect(row.find('.status-terms').text()).toEqual('HeLa-S3');
         });
 
-        it('has a good summary panel', function() {
-            var panel = TestUtils.findRenderedDOMComponentWithClass(antibody, 'data-display').getDOMNode();
+        test('has a good summary panel', () => {
+            const panel = antibody.find('.data-display');
 
-            var item = panel.querySelector('[data-test="source"]');
-            var itemDescription = item.getElementsByTagName('dd')[0];
-            var anchor = itemDescription.getElementsByTagName('a')[0];
-            expect(itemDescription.textContent).toEqual('Aviva');
-            expect(anchor.getAttribute('href')).toEqual('http://www.avivasysbio.com');
+            let item = panel.find('[data-test="source"]');
+            let itemDescription = item.find('dd');
+            let anchor = itemDescription.at(0).find('a');
+            expect(itemDescription.text()).toEqual('Aviva');
+            expect(anchor.prop('href')).toEqual('http://www.avivasysbio.com');
 
-            item = panel.querySelector('[data-test="productid"]');
-            itemDescription = item.getElementsByTagName('dd')[0];
-            anchor = itemDescription.getElementsByTagName('a')[0];
-            expect(itemDescription.textContent).toEqual('ARP40383_T100');
-            expect(anchor.getAttribute('href')).toEqual('http://www.avivasysbio.com/anti-hnrpa1-antibody-n-terminal-region-arp40383-t100.html');
+            item = panel.find('[data-test="productid"]');
+            itemDescription = item.find('dd');
+            anchor = itemDescription.find('a');
+            expect(itemDescription.text()).toEqual('ARP40383_T100');
+            expect(anchor.at(0).prop('href')).toEqual('http://www.avivasysbio.com/anti-hnrpa1-antibody-n-terminal-region-arp40383-t100.html');
 
-            item = panel.querySelector('[data-test="lotid"]');
-            itemDescription = item.getElementsByTagName('dd')[0];
-            expect(itemDescription.textContent).toEqual('QC9473-091124');
+            item = panel.find('[data-test="lotid"]');
+            itemDescription = item.find('dd');
+            expect(itemDescription.text()).toEqual('QC9473-091124');
 
-            item = panel.querySelector('[data-test="targets"]');
-            itemDescription = item.getElementsByTagName('dd')[0];
-            anchor = itemDescription.getElementsByTagName('a')[0];
-            expect(itemDescription.textContent).toEqual('HNRNPA1 (Homo sapiens)');
-            expect(anchor.getAttribute('href')).toEqual('/targets/HNRNPA1-human/');
+            item = panel.find('[data-test="targets"]');
+            itemDescription = item.find('dd');
+            anchor = itemDescription.find('a');
+            expect(itemDescription.text()).toEqual('HNRNPA1 (Homo sapiens)');
+            expect(anchor.at(0).prop('href')).toEqual('/targets/HNRNPA1-human/');
 
-            item = panel.querySelector('[data-test="host"]');
-            itemDescription = item.getElementsByTagName('dd')[0];
-            expect(itemDescription.textContent).toEqual('rabbit');
+            item = panel.find('[data-test="host"]');
+            itemDescription = item.find('dd');
+            expect(itemDescription.text()).toEqual('rabbit');
 
-            item = panel.querySelector('[data-test="clonality"]');
-            itemDescription = item.getElementsByTagName('dd')[0];
-            expect(itemDescription.textContent).toEqual('polyclonal');
+            item = panel.find('[data-test="clonality"]');
+            itemDescription = item.find('dd');
+            expect(itemDescription.text()).toEqual('polyclonal');
 
-            item = panel.querySelector('[data-test="isotype"]');
-            itemDescription = item.getElementsByTagName('dd')[0];
-            expect(itemDescription.textContent).toEqual('IgG');
+            item = panel.find('[data-test="isotype"]');
+            itemDescription = item.find('dd');
+            expect(itemDescription.text()).toEqual('IgG');
 
-            item = panel.querySelector('[data-test="antigensequence"]');
-            itemDescription = item.getElementsByTagName('dd')[0];
-            expect(itemDescription.textContent).toEqual('MSKSESPKEPEQLRKLFIGGLSFETTDESLRSHFEQWGTLTDCVVMRDPN');
+            item = panel.find('[data-test="antigensequence"]');
+            itemDescription = item.find('dd');
+            expect(itemDescription.text()).toEqual('MSKSESPKEPEQLRKLFIGGLSFETTDESLRSHFEQWGTLTDCVVMRDPN');
         });
     });
 });
