@@ -1,13 +1,9 @@
 import DataColors from '../datacolors';
 
-jest.autoMockOff();
-jest.unmock('../datacolors');
 
-describe('DataColor Module', function() {
-    var React, TestUtils, assembleGraph, graphException, context, _, collectNodes;
-
-    describe('Returns correct colors for keys', function() {
-        var testKeys = [
+describe('DataColor Module', () => {
+    describe('Returns correct colors for keys', () => {
+        const testKeys = [
             'immortalized cell line',
             'tissue',
             'primary cell',
@@ -15,28 +11,28 @@ describe('DataColor Module', function() {
             'stem cell',
             'in vitro differentiated cells',
             'induced pluripotent stem cell line',
-            'secondary cell'
+            'secondary cell',
         ];
-        var dataColorsInstance = new DataColors(testKeys);
+        const dataColorsInstance = new DataColors(testKeys);
 
-        it('Returns correct colors for small array', function() {
-            let testColors = dataColorsInstance.colorList(['stem cell', 'primary cell']);
+        it('Returns correct colors for small array', () => {
+            const testColors = dataColorsInstance.colorList(['stem cell', 'primary cell']);
             expect(testColors[0]).toEqual('#9b009b');
             expect(testColors[1]).toEqual('#ff9a00');
         });
 
-        it('Returns medium gray for a non-existent key', function() {
-            let testColors = dataColorsInstance.colorList(['stem DNA']);
+        it('Returns medium gray for a non-existent key', () => {
+            const testColors = dataColorsInstance.colorList(['stem DNA']);
             expect(testColors[0]).toEqual('#808080');
         });
 
-        it('Wraps around the used color if more keys than colors', function() {
-            let testColors = dataColorsInstance.colorList(['secondary cell']);
+        it('Wraps around the used color if more keys than colors', () => {
+            const testColors = dataColorsInstance.colorList(['secondary cell']);
             expect(testColors[0]).toEqual('#2f62cf');
         });
 
-        it('Wraps around the used color if more keys than colors', function() {
-            let testColors = dataColorsInstance.colorList(['primary cell'], {shade: 10});
+        it('Wraps around the used color if more keys than colors', () => {
+            const testColors = dataColorsInstance.colorList(['primary cell'], { shade: 10 });
             expect(testColors[0]).toEqual('#ffb41a');
         });
     });
