@@ -1,192 +1,164 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import _ from 'underscore';
+import url from 'url';
 
-'use strict';
-
-jest.autoMockOff();
-
-// Fixes https://github.com/facebook/jest/issues/78
-jest.dontMock('react');
-jest.dontMock('underscore');
+// Import test component and data.
+import { Experiment } from '../experiment';
+import context from '../testdata/experiment';
 
 
-describe('Experiment Page', function() {
-    var React, TestUtils, Experiment, FetchContext, context, _;
+describe('Experiment Page', () => {
+    describe('Minimal Experiment', () => {
+        let experiment;
+        let summary;
 
-    beforeEach(function() {
-        React = require('react');
-        TestUtils = require('react/lib/ReactTestUtils');
-        _ = require('underscore');
+        beforeAll(() => {
+            const contextMin = _.clone(context);
+            contextMin.references = [require('../testdata/publication/PMID16395128'), require('../testdata/publication/PMID23000965')];
+            experiment = mount(
+                <Experiment context={context} />
+            );
 
-        Experiment = require('../experiment').Experiment;
-        context = require('../testdata/experiment');
-
-        FetchContext = {
-            fetch: function(url, options) {
-                return Promise.resolve({json: () => ({'@graph': []})});
-            }
-        };
-    });
-
-    describe('Minimal Experiment', function() {
-        var experiment, summary, defTerms, defDescs;
-
-        beforeEach(function() {
-            context.references = [require('../testdata/publication/PMID16395128'), require('../testdata/publication/PMID23000965')];
-            experiment = React.withContext(FetchContext, function() {
-                return TestUtils.renderIntoDocument(
-                    <Experiment context={context} />
-                );
-            });
-
-            summary = TestUtils.scryRenderedDOMComponentsWithClass(experiment, 'data-display');
-            defTerms = summary[0].getDOMNode().getElementsByTagName('dt');
-            defDescs = summary[0].getDOMNode().getElementsByTagName('dd');
+            summary = experiment.find('.data-display');
         });
 
-        it('has proper links in dbxrefs key-value', function() {
-            var item = summary[0].getDOMNode().querySelector('[data-test="external-resources"]');
-            var desc = item.getElementsByTagName('dd')[0];
-            var dbxrefs = desc.getElementsByTagName('a');
-            expect(dbxrefs.length).toEqual(2);
-            expect(dbxrefs[0].getAttribute('href')).toEqual('http://genome.ucsc.edu/cgi-bin/hgTracks?tsCurTab=advancedTab&tsGroup=Any&tsType=Any&hgt_mdbVar1=dccAccession&hgt_tSearch=search&hgt_tsDelRow=&hgt_tsAddRow=&hgt_tsPage=&tsSimple=&tsName=&tsDescr=&db=hg19&hgt_mdbVal1=wgEncodeEH003317');
-            expect(dbxrefs[1].getAttribute('href')).toEqual('https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM1010811');
+        test('has proper links in dbxrefs key-value', () => {
+            const item = summary.at(0).find('[data-test="external-resources"]');
+            const desc = item.find('dd');
+            const dbxrefs = desc.at(0).find('a');
+            expect(dbxrefs).toHaveLength(2);
+            expect(dbxrefs.at(0).prop('href')).toEqual('http://genome.ucsc.edu/cgi-bin/hgTracks?tsCurTab=advancedTab&tsGroup=Any&tsType=Any&hgt_mdbVar1=dccAccession&hgt_tSearch=search&hgt_tsDelRow=&hgt_tsAddRow=&hgt_tsPage=&tsSimple=&tsName=&tsDescr=&db=hg19&hgt_mdbVal1=wgEncodeEH003317');
+            expect(dbxrefs.at(1).prop('href')).toEqual('https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM1010811');
         });
 
-        it('has proper release date', function() {
-            var item = summary[0].getDOMNode().querySelector('[data-test="date-released"]');
-            var desc = item.getElementsByTagName('dd')[0];
-            expect(desc.textContent).toEqual('2011-10-29');
+        test('has proper release date', () => {
+            const item = summary.at(0).find('[data-test="date-released"]');
+            const desc = item.find('dd');
+            expect(desc.at(0).text()).toEqual('2011-10-29');
         });
 
-        it('has one experiment status element in header', function() {
-            var statusList = TestUtils.findRenderedDOMComponentWithClass(experiment, 'status-list').getDOMNode();
-            expect(statusList.hasChildNodes()).toBeTruthy();
-            expect(statusList.childNodes.length).toEqual(1);
+        test('has one experiment status element in header', () => {
+            const statusList = experiment.find('.status-list');
+            expect(statusList.children().exists()).toBeTruthy();
+            expect(statusList.children()).toHaveLength(1);
         });
     });
 
-    describe('Document Panel', function() {
-        var experiment, doc;
+    describe('Document Panel', () => {
+        let experiment;
+        let doc;
 
-        beforeEach(function() {
-            require('../biosample.js').Document;
-            var context_doc = _.clone(context);
-            context_doc.documents = [require('../testdata/document/myerschipseq')];
-            experiment = React.withContext(FetchContext, function() {
-                return TestUtils.renderIntoDocument(
-                    <Experiment context={context_doc} />
-                );
-            });
-            doc = TestUtils.findRenderedDOMComponentWithClass(experiment, 'type-Document').getDOMNode();;
+        beforeAll(() => {
+            const contextDoc = _.clone(context);
+            contextDoc.documents = [require('../testdata/document/myerschipseq')];
+            experiment = mount(
+                <Experiment context={contextDoc} />
+            );
+            doc = experiment.find('.type-Document');
         });
 
-        it('has one document panel with a PDF image anchor', function() {
-            var figures = doc.getElementsByTagName('figure');
-            expect(figures.length).toEqual(1);
-            var anchors = figures[0].getElementsByClassName('file-pdf');
-            expect(anchors.length).toEqual(1);
+        test('has one document panel with a PDF image anchor', () => {
+            const figures = doc.find('figure');
+            expect(figures).toHaveLength(1);
+            const anchors = figures.find('.file-pdf');
+            expect(anchors).toHaveLength(1);
         });
 
-        it('has a single proper download link', function() {
-            var url = require('url');
-
-            var dlBar = doc.getElementsByClassName('document__file');
-            expect(dlBar.length).toEqual(1);
-            var anchors = dlBar[0].getElementsByTagName('a');
-            expect(anchors.length).toEqual(2);
-            expect(url.parse(anchors[0].getAttribute('href')).pathname).toEqual('/documents/df9dd0ec-c1cf-4391-a745-a933ab1af7a7/@@download/attachment/Myers_Lab_ChIP-seq_Protocol_v042211.pdf');
+        test('has a single proper download link', () => {
+            const dlBar = doc.find('.document__file');
+            expect(dlBar).toHaveLength(1);
+            const anchors = dlBar.find('a');
+            expect(anchors).toHaveLength(2);
+            expect(url.parse(anchors.at(0).prop('href')).pathname).toEqual('/documents/df9dd0ec-c1cf-4391-a745-a933ab1af7a7/@@download/attachment/Myers_Lab_ChIP-seq_Protocol_v042211.pdf');
         });
 
-        it('has two key-value pairs, and proper DL link', function() {
-            var trigger = doc.getElementsByClassName('document__file-detail-switch');
-            TestUtils.Simulate.click(trigger[0]);
+        test('has two key-value pairs, and proper DL link', () => {
+            const trigger = doc.find('.document__file-detail-switch');
+            trigger.at(0).simulate('click');
 
-            var docKeyValue = doc.getElementsByClassName('key-value-doc');
-            expect(docKeyValue.length).toEqual(1);
-            var defTerms = docKeyValue[0].getElementsByTagName('dt');
-            expect(defTerms.length).toEqual(3);
-            var defDescs = docKeyValue[0].getElementsByTagName('dd');
-            expect(defDescs.length).toEqual(3);
+            const docKeyValue = doc.find('.key-value-doc');
+            expect(docKeyValue).toHaveLength(1);
+            const defTerms = docKeyValue.at(0).find('dt');
+            expect(defTerms).toHaveLength(3);
+            const defDescs = docKeyValue.at(0).find('dd');
+            expect(defDescs).toHaveLength(3);
         });
     });
 
-    describe('Document Panel References', function() {
-        var experiment, doc;
+    describe('Document Panel References', () => {
+        let experiment;
+        let doc;
 
-        beforeEach(function() {
-            require('../biosample.js').Document;
-            var context_doc = _.clone(context);
-            context_doc.documents = [require('../testdata/document/wgEncodeSydhHist-refs')];
-            experiment = React.withContext(FetchContext, function() {
-                return TestUtils.renderIntoDocument(
-                    <Experiment context={context_doc} />
-                );
-            });
-            doc = TestUtils.findRenderedDOMComponentWithClass(experiment, 'type-Document').getDOMNode();
+        beforeAll(() => {
+            const contextDoc = _.clone(context);
+            contextDoc.documents = [require('../testdata/document/wgEncodeSydhHist-refs')];
+            experiment = mount(
+                <Experiment context={contextDoc} />
+            );
+            doc = experiment.find('.type-Document');
         });
 
-        it('has five key-value pairs, and two good references links', function() {
-            var url = require('url');
-            var docKeyValue = doc.getElementsByClassName('key-value-doc');
-            expect(docKeyValue.length).toEqual(1);
-            var defTerms = docKeyValue[0].getElementsByTagName('dt');
-            expect(defTerms.length).toEqual(3);
-            var defDescs = docKeyValue[0].getElementsByTagName('dd');
-            expect(defDescs.length).toEqual(3);
+        test('has five key-value pairs, and two good references links', () => {
+            const docKeyValue = doc.find('.key-value-doc');
+            expect(docKeyValue).toHaveLength(1);
+            const defTerms = docKeyValue.at(0).find('dt');
+            expect(defTerms).toHaveLength(3);
+            const defDescs = docKeyValue.at(0).find('dd');
+            expect(defDescs).toHaveLength(3);
         });
     });
 
-    describe('Replicate Panels', function() {
-        var experiment, replicates, summary;
+    describe('Replicate Panels', () => {
+        let experiment;
+        let summary;
 
-        beforeEach(function() {
-            var context_rep = _.clone(context);
-            context_rep.replicates = [require('../testdata/replicate/human'), require('../testdata/replicate/mouse')];
-            context_rep.replicates[0].library = require('../testdata/library/sid38806');
-            context_rep.replicates[1].library = require('../testdata/library/sid38807');
-            context_rep.files = [require('../testdata/file/fastq')];
-            experiment = React.withContext(FetchContext, function() {
-                return TestUtils.renderIntoDocument(
-                    <Experiment context={context_rep} />
-                );
-            });
-            summary = TestUtils.scryRenderedDOMComponentsWithClass(experiment, 'data-display');
+        beforeAll(() => {
+            const contextRep = _.clone(context);
+            contextRep.replicates = [require('../testdata/replicate/human'), require('../testdata/replicate/mouse')];
+            contextRep.replicates[0].library = require('../testdata/library/sid38806');
+            contextRep.replicates[1].library = require('../testdata/library/sid38807');
+            contextRep.files = [require('../testdata/file/fastq')];
+            experiment = mount(
+                <Experiment context={contextRep} />
+            );
+            summary = experiment.find('.data-display');
         });
 
-        it('has proper treatment', function() {
-            var item = summary[0].getDOMNode().querySelector('[data-test="treatments"]');
-            var desc = item.getElementsByTagName('dd')[0];
-            expect(desc.textContent).toEqual('97.2 nM doxycycline hyclate (CHEBI:34730) for 6 hour [1-1]');
+        test('has proper treatment', () => {
+            const item = summary.find('[data-test="treatments"]');
+            const desc = item.find('dd');
+            expect(desc.text()).toEqual('97.2 nM doxycycline hyclate (CHEBI:34730) for 6 hour [1-1]');
         });
 
-        it('has proper strand specificity', function() {
-            var item = summary[0].getDOMNode().querySelector('[data-test="strandspecificity"]');
-            var desc = item.getElementsByTagName('dd')[0];
-            expect(desc.textContent).toEqual('Strand-specific');
+        test('has proper strand specificity', () => {
+            const item = summary.find('[data-test="strandspecificity"]');
+            const desc = item.find('dd');
+            expect(desc.text()).toEqual('Strand-specific');
         });
 
-        it('has proper spikeins', function() {
-            var item = summary[0].getDOMNode().querySelector('[data-test="spikeins"]');
-            var desc = item.getElementsByTagName('dd')[0];
-            expect(desc.textContent).toEqual('ENCSR000AJW [1-1]');
+        test('has proper spikeins', () => {
+            const item = summary.find('[data-test="spikeins"]');
+            const desc = item.find('dd');
+            expect(desc.text()).toEqual('ENCSR000AJW [1-1]');
         });
     });
 
-    describe('Alternate accession display', function() {
-        var experiment, alt;
+    describe('Alternate accession display', () => {
+        let experiment;
+        let alt;
 
-        beforeEach(function() {
-            var context_alt = _.clone(context);
-            context_alt.alternate_accessions = ["ENCSR000ACT", "ENCSR999NOF"];
-            experiment = React.withContext(FetchContext, function() {
-                return TestUtils.renderIntoDocument(
-                    <Experiment context={context_alt} />
-                );
-            });
-            alt = TestUtils.findRenderedDOMComponentWithClass(experiment, 'repl-acc');
+        beforeAll(() => {
+            const contextAlt = _.clone(context);
+            contextAlt.alternate_accessions = ['ENCSR000ACT', 'ENCSR999NOF'];
+            experiment = mount(
+                <Experiment context={contextAlt} />
+            );
+            alt = experiment.find('.repl-acc');
         });
 
-        it('displays two alternate accessions', function() {
-            expect(alt.getDOMNode().textContent).toEqual('Replaces ENCSR000ACT, ENCSR999NOF');
+        test('displays two alternate accessions', () => {
+            expect(alt.text()).toEqual('Replaces ENCSR000ACT, ENCSR999NOF');
         });
     });
 });
