@@ -1,21 +1,17 @@
 'use strict';
 var React = require('react');
+import createReactClass from 'create-react-class';
+import { auditDecor } from './audit';
 var globals = require('./globals');
 var navigation = require('./navigation');
 var dbxref = require('./dbxref');
 var search = require('./search');
-var audit = require('./audit');
 
 var Breadcrumbs = navigation.Breadcrumbs;
 var DbxrefList = dbxref.DbxrefList;
-var AuditIndicators = audit.AuditIndicators;
-var AuditDetail = audit.AuditDetail;
-var AuditMixin = audit.AuditMixin;
 
 
-var Publication = module.exports.Panel = React.createClass({
-    mixins: [AuditMixin],
-
+var PublicationComponent = createReactClass({
     render: function() {
         var context = this.props.context;
         var itemClass = globals.itemClass(context, 'view-item');
@@ -34,8 +30,8 @@ var Publication = module.exports.Panel = React.createClass({
             <div className={itemClass}>
                 <Breadcrumbs root='/search/?type=publication' crumbs={crumbs} />
                 <h2>{context.title}</h2>
-                <AuditIndicators audits={context.audit} id="publication-audit" />
-                <AuditDetail audits={context.audit} except={context['@id']} id="publication-audit" />
+                {this.props.auditIndicators(context.audit, 'publication-audit')}
+                {this.props.auditDetail(context.audit, 'publication-audit', { except: context['@id'] })}
                 {context.authors ? <div className="authors">{context.authors}.</div> : null}
                 <div className="journal">
                     <Citation {...this.props} />
@@ -62,10 +58,13 @@ var Publication = module.exports.Panel = React.createClass({
     }
 });
 
+const Publication = auditDecor(PublicationComponent);
+module.exports.Publication = Publication; // Only need to export for Jest tests
+
 globals.content_views.register(Publication, 'Publication');
 
 
-var Citation = module.exports.Citation = React.createClass({
+var Citation = module.exports.Citation = createReactClass({
     render: function() {
         var context = this.props.context;
         return (
@@ -78,7 +77,7 @@ var Citation = module.exports.Citation = React.createClass({
 });
 
 
-var Abstract = React.createClass({
+var Abstract = createReactClass({
     render: function() {
         var context = this.props.context;
         return (
@@ -125,13 +124,12 @@ var Abstract = React.createClass({
 });
 
 
-var SupplementaryData = React.createClass({
+var SupplementaryData = createReactClass({
     render: function() {
         var data = this.props.data;
         return (
-            <section>
-                {this.props.key > 0 ? <hr /> : null}
-                <dl className="key-value" key={this.props.key}>
+            <section className="supplementary-data">
+                <dl className="key-value">
                     {data.supplementary_data_type ?
                         <div data-test="availabledata">
                             <dt>Available data</dt>
@@ -166,7 +164,7 @@ var SupplementaryData = React.createClass({
 });
 
 
-var SupplementaryDataListing = React.createClass({
+var SupplementaryDataListing = createReactClass({
     getInitialState: function() {
         return {excerptExpanded: false};
     },
@@ -215,8 +213,7 @@ var SupplementaryDataListing = React.createClass({
 });
 
 
-var Listing = React.createClass({
-    mixins: [search.PickerActionsMixin, AuditMixin],
+var ListingComponent = createReactClass({
     render: function() {
         var result = this.props.context;
         var authorList = result.authors && result.authors.length ? result.authors.split(', ', 4) : [];
@@ -225,11 +222,11 @@ var Listing = React.createClass({
         return (
             <li>
                 <div className="clearfix">
-                    {this.renderActions()}
+                    <search.PickerActions {...this.props} />
                     <div className="pull-right search-meta">
                         <p className="type meta-title">Publication</p>
                         <p className="type meta-status">{' ' + result.status}</p>
-                        <AuditIndicators audits={result.audit} id={this.props.context['@id']} search />
+                        {this.props.auditIndicators(result.audit, result['@id'], { search: true })}
                     </div>
                     <div className="accession"><a href={result['@id']}>{result.title}</a></div>
                     <div className="data-row">
@@ -245,10 +242,12 @@ var Listing = React.createClass({
                         : null}
                     </div>
                 </div>
-                <AuditDetail audits={result.audit} except={result['@id']} id={this.props.context['@id']} forcedEditLink />
+                {this.props.auditDetail(result.audit, result['@id'], { forcedEditLink: true })}
             </li>
         );
     }
 });
+
+const Listing = auditDecor(ListingComponent);
 
 globals.listing_views.register(Listing, 'Publication');
