@@ -1,9 +1,10 @@
 'use strict';
 var React = require('react');
+import createReactClass from 'create-react-class';
 var globals = require('./globals');
 
 
-var TestingRenderErrorPanel = module.exports.TestingRenderErrorPanel = React.createClass({
+var TestingRenderErrorPanel = module.exports.TestingRenderErrorPanel = createReactClass({
     render: function() {
         console.log('log');
         console.warn('warn');
