@@ -1,46 +1,39 @@
-'use strict';
+import Registry from '../registry';
 
-jest.dontMock('../registry');
-jest.dontMock('underscore');
-var Registry = require('../registry');
+const testObj = { '@type': ['Test', 'Item'] };
+const specificObj = { '@type': ['Specific', 'Item'] };
+const otherObj = { '@type': ['Other'] };
 
-var test_obj = {'@type': ['Test', 'Item']};
-var specific_obj = {'@type': ['Specific', 'Item']};
-var other_obj = {'@type': ['Other']};
-
-var views = [
-    {for_: 'Item'},
-    {for_: 'Specific'},
-    {name: 'named', for_: 'Item'}
+const views = [
+    { for_: 'Item' },
+    { for_: 'Specific' },
+    { name: 'named', for_: 'Item' },
 ];
 
-var make_one = function () {
-    var registry = new Registry();
-    views.forEach(function (view) {
+/* eslint no-underscore-dangle: ["error", { "allow": ["for_"] }]*/
+function makeOne() {
+    const registry = new Registry();
+    views.forEach((view) => {
         registry.register(view, view.for_, view.name);
     });
-    registry.fallback = function () {
-        return {fallback: true};
-    };
+    registry.fallback = () => ({ fallback: true });
     return registry;
-};
+}
 
-describe("The registry library", function() {
-
-    it("is able to lookup views for item in order of specificity", function() {
-        var registry = make_one();
-        expect(registry.lookup(test_obj).for_).toBe('Item');
-        expect(registry.lookup(specific_obj).for_).toBe('Specific');
+describe('The registry library', () => {
+    it('is able to lookup views for item in order of specificity', () => {
+        const registry = makeOne();
+        expect(registry.lookup(testObj).for_).toBe('Item');
+        expect(registry.lookup(specificObj).for_).toBe('Specific');
     });
 
-    it("is able to lookup named views for item", function() {
-        var registry = make_one();
-        expect(registry.lookup(test_obj, 'named').name).toBe('named');
+    it('is able to lookup named views for item', () => {
+        const registry = makeOne();
+        expect(registry.lookup(testObj, 'named').name).toBe('named');
     });
 
-    it("is able to fallback to view for objects with unknown types", function() {
-        var registry = make_one();
-        expect(registry.lookup(other_obj).fallback).toBe(true);
+    it('is able to fallback to view for objects with unknown types', () => {
+        const registry = makeOne();
+        expect(registry.lookup(otherObj).fallback).toBe(true);
     });
-
 });
